separating entry_points from helpers
