+++ docs/_sources/entrypoints.rst.txt
@@ -1,0 +1,11 @@
+:py:data:`rsudp.entrypoints` (convenience functions)
+=====================================================
+
+These are some convenience functions for editing rsudp's settings and
+monitoring log output.
+
+.. automodule:: rsudp.entrypoints
+    :members:
+
+
+`Back to top ↑ <#top>`_
