+++ docs/_sources/index.rst.txt
@@ -74,6 +74,7 @@ The modules available in rsudp are organized by type below.
     init
     raspberryshake
     helpers
+    entrypoints
 
 .. toctree::
     :maxdepth: 2
