+++ docs/client.html
@@ -37,7 +37,7 @@
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
     <link rel="next" title="rsudp.packetloss (track packets)" href="packetloss.html" />
-    <link rel="prev" title="rsudp.helpers (helper functions)" href="helpers.html" /> 
+    <link rel="prev" title="rsudp.entrypoints (convenience functions)" href="entrypoints.html" /> 
 </head>
 
 <body class="wy-body-for-nav">
@@ -114,6 +114,7 @@
 <li class="toctree-l1"><a class="reference internal" href="init.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.__init__</span></code> (initialization functions)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="raspberryshake.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.raspberryshake</span></code> (main library)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="helpers.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.helpers</span></code> (helper functions)</a></li>
+<li class="toctree-l1"><a class="reference internal" href="entrypoints.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.entrypoints</span></code> (convenience functions)</a></li>
 </ul>
 <p class="caption"><span class="caption-text">Clients</span></p>
 <ul class="current">
@@ -317,7 +318,7 @@ testing-specific actions should be taken in routines.</p>
         <a href="packetloss.html" class="btn btn-neutral float-right" title="rsudp.packetloss (track packets)" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
       
       
-        <a href="helpers.html" class="btn btn-neutral float-left" title="rsudp.helpers (helper functions)" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
+        <a href="entrypoints.html" class="btn btn-neutral float-left" title="rsudp.entrypoints (convenience functions)" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
       
     </div>
