+++ docs/daemon.html
@@ -119,6 +119,7 @@
 <li class="toctree-l1"><a class="reference internal" href="init.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.__init__</span></code> (initialization functions)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="raspberryshake.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.raspberryshake</span></code> (main library)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="helpers.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.helpers</span></code> (helper functions)</a></li>
+<li class="toctree-l1"><a class="reference internal" href="entrypoints.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.entrypoints</span></code> (convenience functions)</a></li>
 </ul>
 <p class="caption"><span class="caption-text">Clients</span></p>
 <ul>
@@ -267,7 +268,7 @@ systemctl --user start rsudp.service
 </div>
 <p>rsudp should start within about 30 seconds.
 You can monitor the program’s output in real time by entering the following command
-(source function: <a class="reference internal" href="helpers.html#rsudp.helpers.ep_tailf_log" title="rsudp.helpers.ep_tailf_log"><code class="xref py py-func docutils literal notranslate"><span class="pre">rsudp.helpers.ep_tailf_log()</span></code></a>):</p>
+(source function: <a class="reference internal" href="entrypoints.html#rsudp.entrypoints.ep_tailf_log" title="rsudp.entrypoints.ep_tailf_log"><code class="xref py py-func docutils literal notranslate"><span class="pre">rsudp.entrypoints.ep_tailf_log()</span></code></a>):</p>
 <div class="highlight-bash notranslate"><div class="highlight"><pre><span></span>rs-tailf
 </pre></div>
 </div>
