+++ docs/genindex.html
@@ -113,6 +113,7 @@
 <li class="toctree-l1"><a class="reference internal" href="init.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.__init__</span></code> (initialization functions)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="raspberryshake.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.raspberryshake</span></code> (main library)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="helpers.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.helpers</span></code> (helper functions)</a></li>
+<li class="toctree-l1"><a class="reference internal" href="entrypoints.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.entrypoints</span></code> (convenience functions)</a></li>
 </ul>
 <p class="caption"><span class="caption-text">Clients</span></p>
 <ul>
@@ -295,13 +296,13 @@
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="c_write.html#rsudp.c_write.Write.elapse">elapse() (rsudp.c_write.Write method)</a>
 </li>
-      <li><a href="helpers.html#rsudp.helpers.ep_cat_log">ep_cat_log() (in module rsudp.helpers)</a>
+      <li><a href="entrypoints.html#rsudp.entrypoints.ep_cat_log">ep_cat_log() (in module rsudp.entrypoints)</a>
 </li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="helpers.html#rsudp.helpers.ep_edit_settings">ep_edit_settings() (in module rsudp.helpers)</a>
+      <li><a href="entrypoints.html#rsudp.entrypoints.ep_edit_settings">ep_edit_settings() (in module rsudp.entrypoints)</a>
 </li>
-      <li><a href="helpers.html#rsudp.helpers.ep_tailf_log">ep_tailf_log() (in module rsudp.helpers)</a>
+      <li><a href="entrypoints.html#rsudp.entrypoints.ep_tailf_log">ep_tailf_log() (in module rsudp.entrypoints)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -462,6 +463,8 @@
         <li><a href="c_write.html#module-rsudp.c_write">rsudp.c_write</a>
 </li>
         <li><a href="client.html#module-rsudp.client">rsudp.client</a>
+</li>
+        <li><a href="entrypoints.html#module-rsudp.entrypoints">rsudp.entrypoints</a>
 </li>
         <li><a href="helpers.html#module-rsudp.helpers">rsudp.helpers</a>
 </li>
@@ -629,6 +632,13 @@
 
       <ul>
         <li><a href="client.html#module-rsudp.client">module</a>
+</li>
+      </ul></li>
+      <li>
+    rsudp.entrypoints
+
+      <ul>
+        <li><a href="entrypoints.html#module-rsudp.entrypoints">module</a>
 </li>
       </ul></li>
   </ul></td>
