+++ docs/helpers.html
@@ -36,7 +36,7 @@
     <link rel="author" title="About these documents" href="about.html" />
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
-    <link rel="next" title="rsudp.client (central module)" href="client.html" />
+    <link rel="next" title="rsudp.entrypoints (convenience functions)" href="entrypoints.html" />
     <link rel="prev" title="rsudp.raspberryshake (main library)" href="raspberryshake.html" /> 
 </head>
 
@@ -114,6 +114,7 @@
 <li class="toctree-l1"><a class="reference internal" href="init.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.__init__</span></code> (initialization functions)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="raspberryshake.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.raspberryshake</span></code> (main library)</a></li>
 <li class="toctree-l1 current"><a class="current reference internal" href="#"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.helpers</span></code> (helper functions)</a></li>
+<li class="toctree-l1"><a class="reference internal" href="entrypoints.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.entrypoints</span></code> (convenience functions)</a></li>
 </ul>
 <p class="caption"><span class="caption-text">Clients</span></p>
 <ul>
@@ -333,59 +334,6 @@ contains a valid <a class="reference external" href="https://docs.obspy.org/pack
 </ul>
 </dd>
 </dl>
-</dd></dl>
-
-<dl class="py function">
-<dt id="rsudp.helpers.ep_cat_log">
-<code class="sig-prename descclassname">rsudp.helpers.</code><code class="sig-name descname">ep_cat_log</code><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#rsudp.helpers.ep_cat_log" title="Permalink to this definition">¶</a></dt>
-<dd><div class="versionadded">
-<p><span class="versionmodified added">New in version 1.0.3.</span></p>
-</div>
-<p>This function uses a posix system’s <code class="docutils literal notranslate"><span class="pre">cat</span></code> command to print messages in
-the rsudp log file. It is provided for convenience only.</p>
-<p>It is accessible via the console command <code class="docutils literal notranslate"><span class="pre">rs-log</span></code> on posix (Linux, MacOS)
-style operating systems.</p>
-</dd></dl>
-
-<dl class="py function">
-<dt id="rsudp.helpers.ep_edit_settings">
-<code class="sig-prename descclassname">rsudp.helpers.</code><code class="sig-name descname">ep_edit_settings</code><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#rsudp.helpers.ep_edit_settings" title="Permalink to this definition">¶</a></dt>
-<dd><div class="versionadded">
-<p><span class="versionmodified added">New in version 1.0.3.</span></p>
-</div>
-<p>This function calls the system’s default text editor to open the settings
-file for editing. It is provided for convenience only.
-Advanced users may prefer to add an alias in place of this function.
-The alias <strong>should</strong> override the entrypoint command set in rsudp’s
-<code class="docutils literal notranslate"><span class="pre">setup.py</span></code>.</p>
-<p>On Linux and MacOS, adding an alias may look like this:</p>
-<div class="highlight-bash notranslate"><div class="highlight"><pre><span></span><span class="c1"># add the alias definition to the aliases file</span>
-<span class="nb">echo</span> <span class="s2">&quot;alias rsudp-settings=&#39;nano ~/.config/rsudp/rsudp_settings.json&#39;&quot;</span> &gt;&gt; .bash_aliases
-<span class="c1"># then reload the console</span>
-bash
-</pre></div>
-</div>
-<p>To add an alias on Windows via the command prompt is much more difficult.</p>
-<div class="admonition note">
-<p class="admonition-title">Note</p>
-<p>This function has been tested on multiple operating systems, but
-because each system’s functionality and defaults may be different,
-proper operation cannot be not guaranteed.</p>
-</div>
-</dd></dl>
-
-<dl class="py function">
-<dt id="rsudp.helpers.ep_tailf_log">
-<code class="sig-prename descclassname">rsudp.helpers.</code><code class="sig-name descname">ep_tailf_log</code><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#rsudp.helpers.ep_tailf_log" title="Permalink to this definition">¶</a></dt>
-<dd><div class="versionadded">
-<p><span class="versionmodified added">New in version 1.0.3.</span></p>
-</div>
-<p>This function uses a the system’s follow command to follow new
-messages added to the log file. It is provided for convenience only.
-is the equivalent of <code class="docutils literal notranslate"><span class="pre">tail</span> <span class="pre">-f</span> <span class="pre">/tmp/rsudp/rsudp.log</span></code> on Linux/MacOS
-and <code class="docutils literal notranslate"><span class="pre">Get-Content</span> <span class="pre">-Path</span> <span class="pre">&quot;C:/tmp/rsudp/rsudp.log&quot;</span> <span class="pre">-Wait</span></code> on Windows.</p>
-<p>It is accessible via the console command <code class="docutils literal notranslate"><span class="pre">rs-tailf</span></code>.</p>
-<p>The function will run until it receives a keyboard interrupt (CTRL+C).</p>
 </dd></dl>
 
 <dl class="py function">
@@ -680,7 +628,7 @@ you could specify</p>
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="client.html" class="btn btn-neutral float-right" title="rsudp.client (central module)" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
+        <a href="entrypoints.html" class="btn btn-neutral float-right" title="rsudp.entrypoints (convenience functions)" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
       
       
         <a href="raspberryshake.html" class="btn btn-neutral float-left" title="rsudp.raspberryshake (main library)" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
