+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["about","c_alert","c_alertsound","c_consumer","c_custom","c_forward","c_plot","c_rsam","c_telegram","c_testing","c_tweet","c_write","client","contributing","daemon","helpers","index","init","installing","p_producer","packetize","packetloss","raspberryshake","running","settings","t_testdata","test","testing","theory","troubleshooting","youtube"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":2,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.intersphinx":1,"sphinx.ext.todo":2,sphinx:56},filenames:["about.rst","c_alert.rst","c_alertsound.rst","c_consumer.rst","c_custom.rst","c_forward.rst","c_plot.rst","c_rsam.rst","c_telegram.rst","c_testing.rst","c_tweet.rst","c_write.rst","client.rst","contributing.rst","daemon.rst","helpers.rst","index.rst","init.rst","installing.rst","p_producer.rst","packetize.rst","packetloss.rst","raspberryshake.rst","running.rst","settings.rst","t_testdata.rst","test.rst","testing.rst","theory.rst","troubleshooting.rst","youtube.rst"],objects:{"rsudp.__init__":{LevelFormatter:[17,1,1,""],add_debug_handler:[17,3,1,""],get_scap_dir:[17,3,1,""],init_dirs:[17,3,1,""],make_colors_friendly:[17,3,1,""],printE:[17,3,1,""],printM:[17,3,1,""],printW:[17,3,1,""],start_logging:[17,3,1,""]},"rsudp.__init__.LevelFormatter":{format:[17,2,1,""]},"rsudp.c_alert":{Alert:[1,1,1,""]},"rsudp.c_alert.Alert":{run:[1,2,1,""]},"rsudp.c_alertsound":{AlertSound:[2,1,1,""]},"rsudp.c_alertsound.AlertSound":{run:[2,2,1,""]},"rsudp.c_consumer":{Consumer:[3,1,1,""]},"rsudp.c_consumer.Consumer":{run:[3,2,1,""]},"rsudp.c_custom":{Custom:[4,1,1,""]},"rsudp.c_custom.Custom":{run:[4,2,1,""]},"rsudp.c_forward":{Forward:[5,1,1,""]},"rsudp.c_forward.Forward":{run:[5,2,1,""]},"rsudp.c_plot":{Plot:[6,1,1,""]},"rsudp.c_plot.Plot":{deconvolve:[6,2,1,""],figloop:[6,2,1,""],getq:[6,2,1,""],handle_close:[6,2,1,""],handle_resize:[6,2,1,""],mainloop:[6,2,1,""],qu:[6,2,1,""],run:[6,2,1,""],savefig:[6,2,1,""],set_sps:[6,2,1,""],setup_plot:[6,2,1,""],update_plot:[6,2,1,""]},"rsudp.c_rsam":{RSAM:[7,1,1,""]},"rsudp.c_rsam.RSAM":{run:[7,2,1,""]},"rsudp.c_telegram":{Telegrammer:[8,1,1,""]},"rsudp.c_telegram.Telegrammer":{run:[8,2,1,""]},"rsudp.c_testing":{Testing:[9,1,1,""]},"rsudp.c_testing.Testing":{run:[9,2,1,""]},"rsudp.c_tweet":{Tweeter:[10,1,1,""]},"rsudp.c_tweet.Tweeter":{run:[10,2,1,""]},"rsudp.c_write":{Write:[11,1,1,""]},"rsudp.c_write.Write":{elapse:[11,2,1,""],getq:[11,2,1,""],run:[11,2,1,""],set_sps:[11,2,1,""],slicestream:[11,2,1,""],write:[11,2,1,""]},"rsudp.client":{handler:[12,3,1,""],main:[12,3,1,""],mk_p:[12,3,1,""],mk_q:[12,3,1,""],run:[12,3,1,""],start:[12,3,1,""],test:[12,3,1,""],test_mode:[12,3,1,""]},"rsudp.helpers":{conn_stats:[15,3,1,""],deconv_acc_inst:[15,3,1,""],deconv_rbm_inst:[15,3,1,""],deconv_vel_inst:[15,3,1,""],deconvolve:[15,3,1,""],default_settings:[15,3,1,""],dump_default:[15,3,1,""],ep_cat_log:[15,3,1,""],ep_edit_settings:[15,3,1,""],ep_tailf_log:[15,3,1,""],fsec:[15,3,1,""],get_msg_path:[15,3,1,""],get_msg_time:[15,3,1,""],lesser_multiple:[15,3,1,""],msg_alarm:[15,3,1,""],msg_imgpath:[15,3,1,""],msg_reset:[15,3,1,""],msg_term:[15,3,1,""],read_settings:[15,3,1,""],resolve_extra_text:[15,3,1,""],set_channels:[15,3,1,""]},"rsudp.p_producer":{Producer:[19,1,1,""]},"rsudp.p_producer.Producer":{run:[19,2,1,""]},"rsudp.packetize":{cutoff_calc:[20,3,1,""],get_samps:[20,3,1,""],main:[20,3,1,""],packetize:[20,3,1,""]},"rsudp.packetloss":{main:[21,3,1,""],printTTLS:[21,3,1,""],run:[21,3,1,""],signal_handler:[21,3,1,""]},"rsudp.raspberryshake":{ConsumerThread:[22,1,1,""],copy:[22,3,1,""],getCHN:[22,3,1,""],getCHNS:[22,3,1,""],getDATA:[22,3,1,""],getSR:[22,3,1,""],getSTREAM:[22,3,1,""],getTIME:[22,3,1,""],getTR:[22,3,1,""],getTTLCHN:[22,3,1,""],get_inventory:[22,3,1,""],get_ip:[22,3,1,""],handler:[22,3,1,""],initRSlib:[22,3,1,""],make_trace:[22,3,1,""],openSOCK:[22,3,1,""],set_params:[22,3,1,""],update_stream:[22,3,1,""]},"rsudp.t_testdata":{TestData:[25,1,1,""]},"rsudp.t_testdata.TestData":{run:[25,2,1,""],send:[25,2,1,""]},"rsudp.test":{cancel_tests:[26,3,1,""],datadir_permissions:[26,3,1,""],is_connected:[26,3,1,""],logdir_permissions:[26,3,1,""],make_test_settings:[26,3,1,""],permissions:[26,3,1,""],ss_permissions:[26,3,1,""]},rsudp:{__init__:[17,0,0,"-"],c_alert:[1,0,0,"-"],c_alertsound:[2,0,0,"-"],c_consumer:[3,0,0,"-"],c_custom:[4,0,0,"-"],c_forward:[5,0,0,"-"],c_plot:[6,0,0,"-"],c_rsam:[7,0,0,"-"],c_telegram:[8,0,0,"-"],c_testing:[9,0,0,"-"],c_tweet:[10,0,0,"-"],c_write:[11,0,0,"-"],client:[12,0,0,"-"],helpers:[15,0,0,"-"],p_producer:[19,0,0,"-"],packetize:[20,0,0,"-"],packetloss:[21,0,0,"-"],raspberryshake:[22,0,0,"-"],t_testdata:[25,0,0,"-"],test:[26,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:function"},terms:{"00z":22,"01t00":15,"10s":14,"15s":17,"16b":[0,24],"1dv7":22,"211704z":28,"21t19":22,"21t20":22,"23t06":28,"246777z":22,"292000z":22,"435214z":28,"532000z":22,"598944z":28,"599z":15,"boolean":[5,24,28],"break":25,"byte":[15,22,28],"case":[18,24,28,29],"class":[1,2,3,4,5,6,7,8,9,10,11,13,15,17,19,22,24,25,28],"default":[1,6,7,12,14,15,16,18,22,23,26,27,28],"final":[14,28,29],"float":[1,7,22,24],"function":[1,4,6,9,11,12,13,14,20,22,23,24,26,28],"import":[0,6,15,20,22,24,26,28],"int":[6,20,21,22,25],"long":[1,15,24],"new":[4,5,6,7,8,9,10,11,12,13,14,15,18,22,24,25,26,27,28,29],"public":[10,29],"return":[6,11,12,15,17,20,21,22,26,28],"short":[1,2,24],"super":28,"throw":24,"true":[1,4,6,7,11,12,15,17,21,22,24,26,28],"try":[15,26],"while":[0,24,28,29],Adding:[13,16,22],And:28,EHE:15,ENE:15,For:[9,12,15,18,22,24,25,27,28,29],One:29,Pis:24,That:[24,27],The:[0,1,3,5,6,10,11,12,13,15,16,17,19,20,21,22,23,24,26,27,28,29],Then:[25,26,28,29],There:[24,28,29],These:[9,15,17,24],Using:[16,28],Will:22,__init__:[16,28],_eq:[],abc:28,abil:[0,27],abl:[0,24,28,29],abort:29,about:[4,13,14,15,16,23,24,27,28,29],abov:[4,23,24,28,29],acc:[7,24],acceler:[15,24],acceleromet:[0,15,24],accept:15,access:[10,15,24],access_secret:[10,24],access_token:[10,24],access_token_secret:10,accommod:24,accompani:0,accordingli:24,account:[10,24],accumul:22,accur:15,achiev:[0,24],across:[0,6,24,29],action:[0,12,27,28],activ:[0,10,14,17,18,20,21,23,24,29],actual:[13,24],adapt:[17,22],add:[0,10,15,17,24,26,28],add_debug_handl:17,added:[13,15,22,28],adding:[15,27],addit:[0,8,10,13,15,17,28],addition:[0,24],addr:5,address:[5,7,15,22,24,25,29],adjust:[24,30],administr:18,advanc:15,advantag:24,adventur:13,after:[0,6,14,22,23,24,28,29],again:24,agenc:[10,24],agg:13,aggress:24,aim:0,alarm:[2,4,5,6,7,8,9,10,13,15,16,22,23,24,26,27,28],alarm_reset:[1,22],alarmsound:16,alert:[0,1,2,3,6,13,15,16,19,26,27,28],alertsound:[2,24,26,28],algorithm:[6,24],alia:15,alias:15,aliv:[9,19,22,25,28],all:[6,11,14,15,19,22,23,24,26,28,29],allow:[6,13,15,16,17,28,29],almost:29,along:28,alpha:13,alreadi:[13,14,17,24,28],also:[15,19,24,25,27,28,29],alt:24,although:[0,24,29],alwai:[13,14,24],amd:18,among:8,amount:24,amplitud:[7,16],anaconda3:18,anaconda:[18,23],analysi:[16,24],ani:[0,6,15,18,23,24,27,28,29],announc:17,anoth:[5,20,24,28,29],answer:[4,17,22,24],anyth:[12,27],anywai:24,api:[8,10,24,28],api_kei:24,api_secret:24,aplot:6,app:[10,13],appear:[24,29],append:[11,12,17,18,24],appli:[7,10,24],applic:0,appropri:[12,15,18,24],approxim:[8,24],apt:24,arbitrari:27,architectur:[16,18,27],archiv:27,arg:[1,7],argument:[20,21,22],armv7:18,armv7l:18,around:29,arriv:[2,6,24,25],artifici:27,ascii:20,asctim:17,ask:[18,29],aspect:0,assign:22,associ:17,assum:[0,24,28,29],atop:24,attach:[13,29],attribut:[17,22],audibl:0,audioseg:2,authent:24,automat:[8,18,22],avail:[0,15,16,18,22,23,24,26,27,30],averag:[1,2,24,29],avoid:[6,15,24],awai:0,axvlin:13,back:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29],backend:[6,13,26],backtick:[13,29],bak:18,bandpass:[1,24],bar:29,base:[15,18,20,24,27,28],bash:[14,15,18,23],bash_alias:15,basic:[24,27],bat:[18,23],batch:[13,23],beauti:0,becaus:[15,22,24],becom:27,bedrock:24,been:[0,15,18,24,29],beep:24,befor:[17,22,24,28],beforehand:24,begin:[6,23,25],behalf:10,behavior:6,being:[22,24,29],below:[0,16,18,24,27,28,29],beneath:[6,24],berryconda3:18,better:24,between:[21,22,24,28,29],beyond:13,bin:14,binari:24,bind:22,bit:27,block:6,blue:29,boaz:16,bonu:24,bool:[1,2,4,5,6,7,8,10,11,12,15,17,21,22,26,28],boom:[15,16,22,24,27,30],boot:14,bot:[8,10],botfath:24,both:[0,5,6,13,14,15,18,22,24,28],box:22,brace:28,brew:24,brief:[0,24],briefli:28,bring:[0,24],broad:16,broadcast:[6,8,24,28],bug:[16,22,29],build:[22,24,28],built:[5,10,24],button:29,buzzer:13,c_alarm:[],c_alert:[13,15,16,24],c_alertoff:26,c_alerton:26,c_alertsound:[16,24,28],c_anotheron:28,c_consum:[1,2,4,5,6,7,8,9,10,11,12,16,19,25,28],c_custom:[13,16,24,26],c_data:[],c_forward:[16,24,26],c_img:[9,26],c_imgpath:[],c_minise:26,c_mymodul:28,c_mytest:28,c_plai:26,c_plot:[15,16,24,26,28],c_print:26,c_printraw:24,c_process:[],c_reset:[],c_rsam:[16,26],c_telegram:[10,15,16,24,26,28],c_telegramimg:26,c_term:[],c_test:16,c_tweet:[16,24,26,28],c_tweetimg:26,c_write:[16,24,26],calcul:[0,1,3,6,15,19,20,22,24],calibr:15,call:[6,10,15,17,20,22,24,28,29],can:[4,5,6,10,13,14,15,16,18,20,22,23,24,27,28],cancel:26,cancel_test:26,cannot:[6,15,22,24,29],canva:6,capabl:[0,24,27,28],caption:28,captur:17,car:0,care:24,carri:17,cast:[0,29],cat:15,cat_log:[],caus:[6,11,18,22,24],centerpiec:0,central:[6,15,16],certain:[4,6,24],cha:[1,5,6,7,11,15],chan:[7,21,24,26],chang:[14,18,19,20,23,24,27,29],channel:[0,1,5,6,7,11,15,18,21,22,24,26],charact:[0,4,8,10,15,24],chart:[27,28],chat:[8,24],chat_id:[8,24],check:[4,14,15,16,18,24,27,29],child:28,chime:24,chn:22,choos:[5,18,23],chosen:[24,27],citizen:[15,22],classroom:27,clear:6,clearli:[13,29],click:[18,23,29],client:[2,16,22,23,27],clone:14,close:[6,12,23,28],closer:24,code:[0,6,14,18,22,26,28,29],codebas:0,codefil:[4,7,24],collect:0,color:[13,17],com:[13,14,16],combin:28,come:[10,22,23,24,29,30],comma:[24,28],command:[12,13,14,15,17,18,20,21,22,24,27],comment:[6,13],common:[22,29],commun:[0,13,24,28,29],comparison:24,compat:6,complex:0,comput:[0,17,18,23,29],conda:[18,20,21,23,27],config:[14,15,18,23,24,29],configur:[12,23,24,29],conform:28,confus:15,conn_stat:15,connect:[14,15,26,29],consecut:[21,22],consist:18,consol:[1,7,15,16],constant:21,constraint:24,construct:[15,28],consult:24,consum:[1,2,4,5,6,7,8,10,11,12,13,15,16,19,20,22,24,25,26],consumer_kei:10,consumer_secret:10,consumerthread:[22,28],contact:[0,24],contain:[0,6,13,15,20,22,24,25],content:[14,15,27],continu:[0,16,24,25,27,28],contribut:[0,16],control:[13,23,24,28],conveni:[15,24],convert:[4,15,20,22,24,27],cool:28,cooper:29,coordin:11,copi:[13,18,22,29],core:[1,6,11,15,20,22,26,27,28],correct:[4,22,24,29],correctli:[14,18,24,28],cost:0,could:[15,18,24,27,28,29],count:[15,22,24],counter:6,coupl:[17,24],cover:[18,23],cpu:22,creat:[13,14,15,16,17,18,22,23,24,27,28],creation:[8,13,14],crust:0,cryptographi:18,csv:[7,24],ctrl:[15,21,23,24,29],current:[9,15,18,22,26,28],custom:[4,13,16,26,27],custom_set:27,customiz:16,cut:15,cutoff_calc:20,d_matplotlib:26,d_pydub:26,daemon:[0,12,16],dai:11,daili:11,data:[0,1,2,3,4,7,8,9,10,15,16,21,22,23,25,26,28,30],data_dir:11,data_fil:25,datacast:[5,16,22,23,29],datadir_permiss:26,date:[11,17],datefmt:17,datetim:[15,17],dead:14,debain:14,debian:24,debug:[1,7,11,12,24,26,28],decim:22,decod:15,deconv:[1,6,7],deconv_acc_inst:15,deconv_rbm_inst:15,deconv_vel_inst:15,deconvolut:15,deconvolv:[6,15,24,26],def:28,default_set:15,defin:24,definit:[15,18],deliv:5,deliveri:29,demand:0,demo:[16,23],demonstr:[16,27],denot:[21,28],depend:[0,10,15,17,18,26,28],depict:27,deprec:24,depth:16,deriv:22,describ:[13,28,29],descript:[14,24,26,28],design:[0,24,25,27,29],desktop:18,destin:[3,5,12,24,28],detail:[0,6,10,22,24],detect:[0,16,23,26,27],detector:24,determin:[1,17,20,25],develop:[0,10,13,16,18,22,24,28],devic:[0,22,29],diagnost:14,diagram:[9,25,27],dict:[12,15,17,26],dictionari:[12,15,17,26,28],differ:[15,17,23,24,27,28,29],difficult:[10,15,29],digit:24,dir:24,directli:[0,22,24,28],directori:[14,17,18,24,26,27,28],disabl:[6,24],disclaim:[0,4,24],discov:27,disk:[11,15,25,27,28],disp:[7,24],displac:24,displai:[0,1,6,11,15,17,24,29],distanc:29,distribut:[0,3,5,10,18,19,24,28],docstr:28,document:[14,27,28],doe:[0,14,15,24,25,27],doesn:[15,18,24],doing:[24,27],don:[26,27,29],done:[14,15,18,20,22,27,28],doorbel:24,doubl:[18,23],down:[10,11,22],download:[18,22,24,26],downward:24,dport:22,draw:6,drawn:6,drop:[24,29],duckduckgo:29,due:[0,29],dump:[15,23,27],dump_default:15,duplic:28,durat:[1,2,6,24,26],dure:[9,16,25,28],dynam:6,each:[3,6,15,19,20,22,24,28,29],earli:[0,6,8,10,24],earlier:28,earth:0,earthquak:[0,16,23,27],easi:[0,13,23,24,28],easier:[10,13,24],easiest:26,easili:[13,22],echo:[15,18],edg:24,edit:[15,18,24,27],edit_set:24,editor:[15,24],educ:0,eew:[0,24],effect:6,effici:[13,28],ehn:15,ehz:[15,22,24,28],either:[5,6,7,20,23,24,28,29],elaps:[6,11,21],els:[12,29],elsewher:[0,14,24],embed:15,empti:24,emsc:10,enabl:[0,6,7,14,24,26,28],encod:22,encourag:28,end:[4,18,24,25,28],endpoint:24,endtest:25,enhanc:13,enn:15,enough:[0,22],ensur:24,enter:[14,18,23,24,27],entir:[14,24],entri:24,entrypoint:15,environ:[18,20,21,23,27],enz:[15,24],ep_cat_log:15,ep_edit_set:15,ep_tailf_log:[14,15],eq_screenshot:[24,26],equal:24,equival:[14,15],errformat:17,error:[0,6,13,14,15,17,22,24,29],especi:[6,24],essenti:[27,28],etc:[13,18],ethernet:29,even:[13,24,27,29],event:[0,6,12,13,15,17,24],event_tim:[6,15],event_time_str:6,eventu:28,everi:[6,7,14,21,24,29],evt:6,exampl:[3,15,17,18,19,21,22,28,29],exceed:[1,24],except:[17,22,28],excess:24,exec:4,execreload:14,execstart:14,execstartpr:14,execstop:14,execut:[3,4,7,14,18,19,22,24,26,28],exist:[17,24,26,28,29],exit:[24,25,28],experi:[24,29],explain:24,explan:18,explicit:[0,24],explor:0,extens:0,extern:29,extra:[10,24],extra_text:[8,10,15,24],extract:22,extrem:[0,24],face:29,fact:0,fail:14,failur:28,fairli:29,fals:[0,1,2,3,4,5,6,7,8,9,10,11,12,15,17,19,20,21,22,24,25,26,28],far:[0,29],fast:[22,24],fdsn:[22,24,26],featur:[0,13,16,23,29],feed:27,feel:[13,16,28],fetch:9,few:24,ffmpeg:[18,24,28],field:[15,18,24],fifo:28,figloop:6,fignam:15,figur:[0,6,15,24,27,28,29],file:[4,5,6,10,11,14,15,17,18,20,23,24,25,26,27,29],filenam:6,filesystem:24,fill:24,filter:[1,24],find:[5,13,16,22,24,27],fine:22,firewal:16,first:[6,10,14,18,23,24,25,28,29],fit:15,fix:[18,29],flag:[1,19,27,28],flavor:24,flow:[9,16,23,25],fmt:17,focu:23,folder:[18,24],follow:[4,14,15,16,18,21,24,26,27,28,29],footstep:24,forc:[23,24],forese:20,forg:18,fork:24,form:[22,28],formal:28,format:[0,1,4,6,7,13,15,17,20,24,25,27,28,29],formatexcept:17,formatt:[16,28],formattim:17,formerli:23,forum:[13,28,29],forward:[7,11,16,23,26,29],found:24,four:[9,24],fraction:15,frame:[12,21,22],free:[8,24,28,29],freez:14,freq:15,frequenc:[20,21,22,24],friendli:17,from:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30],front:24,fsec:15,full:[6,15],fulli:[0,22,24],fullscreen:[6,24],fun:13,further:[0,24],futur:[18,20,24],fwaddr:[7,24],fwd_alarm:[5,24],fwd_data:[5,24],fwformat:[7,24],fwport:[7,24],gather:28,gempa:22,gener:[10,16,18,27,28,29],geophon:[0,6,15,24],get:[5,6,13,14,15,16,22,24,26,27,28,29],get_inventori:22,get_ip:22,get_msg_path:15,get_msg_tim:15,get_samp:20,get_scap_dir:17,getchn:22,getdata:22,getmessag:17,getq:[6,11,28],getsr:22,getstream:22,gettim:22,gettr:22,getttlchn:22,git:[14,18,23],github:[13,14,16,18,29],give:[10,24,29],given:[15,28,29],glad:29,global:[12,22],globe:29,goe:[24,28,29],going:[0,29],gone:10,got:28,gpio:13,gracefulli:19,grade:[0,24],graphic:14,grasp:0,grav:7,great:29,green:14,grei:14,ground:[3,19],group:24,guarante:[0,15,24,29],gui:[0,6,24],guid:[16,24],guidanc:12,hack:22,had:24,handl:[6,12,23],handle_clos:6,handle_res:6,handler:[12,17,21,22],happen:[6,24,28,29],hard:29,hardcod:20,harmless:18,has:[0,4,6,9,10,11,15,18,21,24,28,29],have:[0,4,6,13,14,15,18,22,24,27,28,29],hdf:[15,22,24],head:18,headless:0,hear:13,heart:6,help:[12,22,24,29],helper:[14,16,22,24],here:[0,1,4,13,17,24,26,28],hierarchi:[9,25,27,28],high:24,higher:24,highli:16,highpass:[1,24,26],hit:18,hobbi:[0,24],hold:24,home:[14,15,17,18,23,24,28],homebrew:24,host:[22,29],hostnam:26,hour:[0,10,21,29],how:[6,13,18,23,24,27,28],howev:[10,18,24,27,29],http:[13,14,16,28,29],hundredth:15,hyperlink:24,ian:16,iannesbitt:[14,15],idea:28,identifi:15,ifttt:13,ignor:[5,11,23,24,28],imag:[6,8,10,15,24,26,28],imgpath:[6,8,9,10,15,26,28],immedi:25,implement:28,impli:[0,24],implic:24,importerror:6,inabl:29,inact:14,includ:[0,10,14,22,23,24,27,28],incom:[1,23],incorpor:15,increas:22,increment:6,indefinit:29,independ:28,index:[1,2,3,4,5,6,7,8,9,10,11,12,19,20,21,22,24,25,26],indic:[12,13,19,29],individu:[27,29],inet:26,inf:20,infinit:28,info:[17,29],infoformat:17,inform:[0,10,15,17,22,24,28,29],ingest:29,init:[17,29],init_dir:17,initd:22,initi:[15,16,21,22,26,28],initrslib:22,input:[20,24],insert:24,inset:28,insid:[24,29],instal:[0,14,16,20,21,23,27,30],instanc:[15,28],instanti:28,instantli:28,instead:[10,27],institut:22,instruct:[4,16,18,24,26,27],instrument:[0,15,16,22,24],integ:[15,18,22],integr:[13,27,28],intend:[0,22,24],interact:13,interfac:17,intern:[24,29],internet:[26,27,29],interrupt:[14,15],interv:[7,24,26],intervent:0,intuit:0,inv:[15,22],inventori:[9,15,22,24,26],involv:13,ioerror:[22,29],is_connect:26,isn:13,isp:29,issu:[0,13,16,24,27,28],item:[24,28],iter:20,its:[0,9,10,14,15,22,24,26,27,28],itself:27,job:29,json:[5,7,15,18,23,24,27,28,29],jump:24,just:[9,11,24,28,29],keep:[0,22,24],kei:[10,18,24,29],keyboard:15,keystrok:21,kid:13,kill:14,kiosk:[0,6,24],know:[16,18,28,29],known:[23,24,29],kwarg:[1,7,22],label:[17,29],laid:28,laptop:18,larg:27,last:[11,24,25,27,28],later:[5,10,18],latest:[18,25],layout:16,lead:24,lean:0,learn:[0,13],least:[15,22,24],leav:[0,18],length:[15,24],less:15,lesser_multipl:15,let:[6,24,28,29],level:17,level_fmt:17,levelformatt:17,liabil:[0,24],libav:[24,28],librari:[6,15,16,23],licens:[0,24],life:[0,24],light:13,like:[3,9,13,14,15,17,18,19,22,23,24,27,28,29],likewis:1,limit:[0,8,10,15,24],line:[4,12,13,14,17,20,21,22,23,24,25,27,28,29],linecach:6,linestyl:13,linewidth:13,linger:14,link:24,linux:[14,15,16],list:[1,3,5,6,11,12,15,19,22,24,28],listen:[1,7,21,22,24,29],lite:[7,24],littl:6,live:[1,7,16],load:[12,24],loc:15,local:[0,16,21,22,24,27],localhost:[22,25],locat:[15,18,22,23,24,29],lock:24,log:[13,14,15,17,22,24,26,29],log_nam:17,logdir:26,logdir_permiss:26,logic:28,loginctl:14,lognam:[],logrecord:17,longer:[6,8,10,15,24],look:[0,10,14,15,16,18,19,22,24,27,28,29],loop:[6,22,25,28],lose:24,loss:[21,29],lost:21,lot:[6,28],low:[0,22],lowpass:[1,24,26],lta:16,lxml:18,mac:[18,24],machin:[14,22,24],maco:[14,15,16],made:[0,11,28],mai:[0,3,6,10,14,15,18,19,23,24,28,29],main:[3,6,12,16,20,21,24,28],mainloop:6,mainpid:14,maintain:[0,28],make:[0,4,6,10,12,13,17,18,20,22,24,27,28,29],make_colors_friendli:17,make_test_set:[26,27,28],make_trac:22,manag:24,mani:[0,6,10,24,28,29],manual:[16,18,24,29],master:[6,12,16,19,26,28],match:[15,24],matplotlib:[6,13,18,26,28],matter:24,max:1,max_len:15,mean:[0,4,18,20,24,27,28],meant:[0,24,27],measur:[7,15,24],media:[0,10,13,28],memori:[0,22],merg:22,messag:[1,2,3,4,5,6,7,8,9,10,11,13,15,16,17,19,20,24,25,26],method:[18,23,24],metric:15,mexican:[8,24],microsecond:15,middl:[16,28],might:[14,24,29],millisecond:[21,22],miniconda3:18,miniconda:18,minimum:0,minise:[11,16,20,26,27],minut:18,misconfigur:14,miss:24,mk_p:[12,28],mk_q:[12,28],mkdir:14,mode:[6,12,16,24,27],model:24,modif:22,modifi:[18,22,23,26,28],modul:[0,4,5,6,8,10,13,16,18,20,22,25,26],monei:0,monitor:[0,13,14,16,19,28,29],month:0,more:[4,5,6,10,13,15,17,22,23,24,27,28,29],most:[0,6,13,22,24,29],motion:[0,3,16,19,24],motor:13,move:[0,24,28,29],mp3:24,mp3file:24,mpl:26,mseed:[20,27],msg:[15,17],msg_alarm:15,msg_imgpath:15,msg_reset:15,msg_term:15,much:[0,15,24,28,29],multi:10,multipl:[0,5,15,24,28],must:[0,6,10,15,18,20,22,24,27,28,29],mymod:28,mymodul:28,n_internet:26,n_inventori:[9,26],n_port:26,nail:10,name:[6,11,15,17,18,21,22,23,24,26,27,28],nano:[14,15],nation:24,nativ:24,natur:0,navig:29,nearbi:24,nearest:15,nearli:0,necessari:[14,15,17,22],necessarili:24,need:[0,13,14,15,16,18,22,23,24,26,27,28,29],neg:[0,24],nesbitt:16,network:[5,14,22,24,25,26],never:24,newbot:24,newlin:[4,24],next:[14,18,23,24,25,28],nodata:22,node:[24,29],noder:24,nois:24,non:[0,24],none:[4,12,15,17],nonetyp:15,normal:[9,27,28],notabl:24,note:[0,27,29],notif:[0,8,10,16],notifi:[1,16,29],now:[6,14,18,22,24,28],nth:6,num:5,number:[6,8,15,20,21,22,24,28],numpi:18,object:[1,2,3,5,6,15,18,19,20,22,28],obspi:[1,3,6,11,15,18,19,20,22,24,27,28],occasion:[0,24],occur:[24,28],odir:17,off:[12,13,15,22,26,28],off_ev:13,offic:30,offici:[0,24],often:[24,29],old:[18,22],on_ev:13,onc:[6,10,24,27,28],one:[0,4,5,13,15,22,23,24,27,28,29],ones:13,onli:[0,8,9,10,12,15,22,23,24,28],onlin:29,open:[0,13,15,18,22,23,24,25,28],opensock:22,oper:[9,11,15,17,18,23,27],operand:17,option:[5,7,10,16,22,24,28],order:[0,5,6,14,15,16,18,22,24,28],ordinari:27,ordinarili:27,org:[13,28,29],organ:[16,28],orig:22,origin:[15,20],oserror:22,other:[0,1,3,5,6,8,11,13,14,16,18,19,23,24,28],otherwis:[7,11,21,26],our:[13,16,18,23,28,29,30],out:[6,15,16,17,22,24,27,28,29],outcom:0,outf:20,outlin:28,output:[0,12,13,14,15,17,20,24,26,29],output_dir:[11,15,24],outsid:29,over:[1,11,18,20,22,27,29],overload:29,overrid:[15,18],overview:[16,24],overwrit:23,overwritten:18,own:[0,13,16,22,24,26],p_data_dir:26,p_log_dir:26,p_log_fil:26,p_log_std:26,p_output_dir:26,p_produc:[1,3,15,16,22,24,28],p_screenshot_dir:26,packag:[20,24],packet:[5,6,7,9,11,15,16,22,24,25,26,27,29],packetloss:[16,29],pad:15,page:[1,2,3,4,5,6,7,8,9,10,11,12,13,16,19,20,21,22,24,25,26,27,29],panama:30,panamaigc:8,param:28,paramet:[1,2,3,4,5,6,7,8,9,10,11,12,15,16,17,19,20,21,22,24,25,26,28],parent:28,pars:[16,22,28],parser:28,part:[0,8,10,15,24,28],partial:15,partwai:18,pass:[2,3,4,5,6,9,12,13,15,16,19,20,21,22,24,25,26],past:[13,14,29],path:[14,15,17,18,23,24,26],patienc:29,paus:18,per:[6,11,15,20,22,23,24],perform:[0,24],perhap:[0,27],period:[0,21,24,29],perman:14,permiss:[10,26,27,29],person:16,piec:[0,24,27],pin:13,ping:24,pip:18,place:[15,20,22,24,28],plai:[13,16,18,26,27,28],plan:28,platform:[0,10,28],playback:24,pleas:[0,4,13,18,24,27,29],plenti:22,plot:[0,3,13,15,16,19,23,26,28],plt:13,plu:29,png:[15,24,28],point:[22,24,28],popul:[3,6,19,24],port:[5,7,15,19,20,21,22,23,24,25,26,27,28,29],portion:24,posit:[0,9,24,25],posix:15,post:[8,10,15,24,28,29],potenti:13,power:[0,6,9,22,24],pre:16,precis:[15,22],prefer:[15,16,23],prepar:25,preparatori:17,press:[23,24,29],pressur:[0,15,16,24],pretti:28,previou:[18,24],print:[7,15,16,17,21,22,26],printdata:[16,26],printfreq:21,printm:[17,22,28],printraw:24,printttl:21,printw:17,prior:[4,22,24,26],privat:29,privileg:18,probabl:[18,28],problem:[13,18,24,27,29],proc:12,proce:[18,24],process:[0,3,6,9,10,12,16,21,23,24,26,27,28],processor:[6,18,24],produc:[1,3,12,13,15,16,20,22,24,27],profession:[0,24],profil:[18,24],program:[0,11,13,14,16,18,21,22,23,24,25,26,27],programmat:28,project:[0,16],prompt:[15,18,23],proper:15,properli:6,properti:[0,24],protect:0,protocol:29,provid:[0,13,14,15,24,29],pull:[18,28],purpos:[11,24,28],put:[6,15,19,24],pydub:[2,26],pyplot:[13,28],python3:20,python:[0,4,15,16,18,20,24,28],qt5:6,quak:24,quarter:6,question:[4,11,13,20,24,28],queue:[1,2,3,4,5,6,7,8,9,10,11,12,13,15,19,24,25,26,28],quickli:[0,8,10],quiet:[7,26],quit:[2,4,6,7,16,19,25,28],r24fa:[15,26,28],r3bcf:22,r940d:22,rais:[1,6,22],ram:0,raspberri:[0,10,14,15,16,18,20,22,24,25,26,27,29,30],raspberryshak:[13,15,16,28,29],raspbian:[14,18,24],raspishak:[13,14,16],rate:[6,10,15,21,22,25,27],ratio:[1,24],raw:[5,15,24],rcb43:22,reach:25,read:[1,2,3,4,5,7,8,10,11,15,16,18,20,22,24,25,27,28],read_set:15,reader:25,readi:[15,18,23,24],readili:24,real:[0,7,14,16],realli:24,reason:[14,18,23,29],reboot:14,rebuild:23,receiv:[15,19,22,24,26,27,28],recept:9,record:[0,6,15,17,27],recur:29,recurs:[1,24],recursive_sta_lta:[1,24],red:[14,28],redraw:6,referenti:22,reflect:[15,29],refresh:[6,24],regard:[0,14],regist:24,reject:24,rel:10,relat:6,releas:[18,24],relev:[24,26,28,29],reli:[0,24,28],reliabl:[10,22],reload:[14,15],remain:0,remind:28,remot:[16,27],remov:24,render:24,replac:18,report:[12,15,21,22,28,29],repositori:[0,18,23,24],repres:9,request:[22,28],requir:[0,4,22,24,27,28],reset:[1,5,9,13,15,22,24,26,28],reset_tim:15,resid:29,resiz:6,resolv:[22,24],resolve_extra_text:15,resourc:[13,16],respect:[24,28],respons:[24,29],rest:[27,28],restart:[0,16,24,29],restartsec:14,restructuredtext:28,result:[12,15,24,26],rewritten:18,richard:16,right:[15,22,28,29],rigor:10,room:0,round:15,router:29,routin:[0,6,11,12,13,27],row:[28,29],rs1d:20,rs4d:[15,24],rsam:[16,26],rsstn:22,rsudp:[13,24,28,30],rsudp_set:[15,18,23,24,29],rtype:28,rule:10,run:[0,1,2,3,5,6,7,8,9,10,11,12,13,14,15,16,19,20,21,22,25,26,27,28,30],runtim:28,s0cde:22,safer:29,sai:[14,24,29],said:24,same:[22,24,25,27,29],sampl:[6,11,15,20,22,27],sasmex:[8,24],save:[0,6,15,17,24,28],save_pct:6,savefig:[6,28],scenario:24,scienc:[0,15,22],scipi:18,screen:[17,24],screencap:6,screenshot:[6,13,15,17,24,26,28],script:[13,14,18,22,23,24,27,29],scroll:29,search:[1,2,3,4,5,6,7,8,9,10,11,12,16,19,20,21,22,24,25,26,29],sec:15,second:[1,2,6,7,11,14,15,18,21,22,24,27,28,29],secret:[10,24],section:[0,5,18,24,28,29],secur:29,sediment:24,see:[0,1,2,4,5,6,7,8,9,10,12,13,14,16,17,19,20,22,23,24,25,27,28,29],seek:25,seem:0,segment:11,seiscomp3:22,seismic:[0,7,16,20,22,27],seismogram:[6,24],seismograph:[0,16],seismolog:0,self:[2,4,6,7,9,11,15,22,25,28],send:[1,6,7,8,10,13,15,17,22,23,24,25,26,27,28,29],send_imag:[8,24],sender:[8,15,17,22,28],sensor:20,sent:[1,2,3,4,5,6,7,8,9,10,11,20,22,24,25,28,29],separ:[6,24],sequenc:[24,28],serial:15,server:[24,26],servic:[6,8,10,14,24,26,29],set:[0,1,4,5,6,7,10,11,12,13,14,15,16,18,21,22,23,25,26,29,30],set_channel:15,set_param:22,set_sp:[6,11],settings_loc:15,setup:[12,15,16,22,29],setup_plot:6,seven:9,sever:[18,23,24,29],shake:[0,6,10,14,15,16,20,22,23,24,25,26,27,30],shorten:24,should:[0,1,12,14,15,18,19,22,24,28,29],show:[0,12,14,27,28],shown:28,shutdown:6,shz:[15,24],side:16,sig:12,sign:24,signal:[1,6,12,21,22,24,28],signal_handl:21,signum:[21,22],similar:[0,13,18,28],similarli:24,simpl:[0,10,11,15,24,25],simpli:[18,24],simultan:[0,24],sinc:[13,21,22,24,28,29],singl:[5,22,24],site:29,size:24,skip:18,sleep:14,slice:11,slicestream:11,slightli:[27,29],slow:29,slower:[6,24],small:[24,27,29],social:[0,10,13,28],socket:[22,25],softwar:[0,13,18,23,24,27,28],solut:22,solv:29,some:[6,13,14,15,18,22,23,24,26,28,29],somehow:14,someon:27,someth:[13,14,16,22,24,28,29],sometim:10,somewher:[24,29],sonar:24,soon:30,sort:28,sound:[0,1,13,16,18,23,26,27,28],soundloc:2,sourc:[0,14,15,23,24,28,29],space:[17,24],span:10,special:16,specif:[1,6,12,15,18,21,22,27,29],specifi:[5,7,15,17,22,24,25,26,27,28,29],spectrogram:[6,24],speed:[24,25],spent:0,sphinx:28,spin:28,sps:[15,22],sqlalchemi:18,ss_permiss:26,sta:[2,16],stabl:0,stackoverflow:[4,17,22,24],stamp:17,stand:28,standard:[0,4,24,28],start:[5,9,11,12,14,16,21,22,24,25,26,27,28],start_log:17,startup:6,stat:15,state:[0,12,13,22,23,24,26],station:[15,22,24,26],stationview:24,statist:21,statu:[14,19,28],stdout:[24,26],steim2:24,step:[17,18,24,28],still:[24,29],stn:22,stomp:24,stop:[12,19,23],store:[22,29],str:[1,4,5,6,7,8,10,11,15,17,20,22,25,26,28],straightforward:29,strang:15,stream:[1,3,6,9,11,15,16,19,20,21,22,23,24,28,30],street:0,strftime:[15,17],strict:29,stricter:10,string:[4,6,15,17,22,24,25,28],structur:[14,26],student:27,stuff:[6,28],style:[4,15,24],sub:[3,12,15,16,22,28],submit:[28,29],subsequ:24,substitut:[0,24],succe:24,success:26,successfulli:24,sudden:[0,16,24],sudo:[14,24],suffer:0,suffic:0,suffici:24,suggest:[13,16],suit:[8,24],summari:27,suppli:[12,24],support:[15,28,29],suppress:7,sure:[4,13,24,27,29],suspici:10,switcher:24,sys:28,system:[0,8,14,15,18,22,23,24],systemctl:[0,16],systemd:14,t_testdata:[16,27],tab:24,tag:13,tail:[14,15],tailf:[14,15],tailf_log:[],take:[6,12,18,24],taken:12,talk:29,target:[14,29],task:[0,3,19],task_don:28,team:[15,29],technic:29,telegram:[0,10,13,15,16,26],telegramm:[8,15,24,28],tell:[18,24,27,28,29],term:[1,2,4,6,7,9,15,19,24,25,26,28],termin:[18,23],test:[0,1,2,3,4,5,6,7,8,10,11,12,14,15,16,17,18,19,20,24],test_mod:12,testb:22,testdata:[9,25,27],testdir:26,text:[10,12,14,15,17,20,24,26,27,29],than:[5,6,10,14,15,18,23,24,27,29],thank:[28,29],thei:[13,17,19,22,23,26,27,28,29],them:[12,15,17,24,25,28,29],theoret:24,theori:[0,28],therefor:16,thi:[0,1,3,4,5,6,8,9,10,11,14,15,16,17,18,20,21,22,23,24,25,26,27,28,29],thing1:28,thing:[0,3,6,8,13,17,19,24,28,29],think:10,thoroughli:24,those:[3,5,20,22],thread:[5,9,12,16,19,22,24,25,27],threadsaf:[6,17],three:[24,28,29],thresh:1,threshold:[1,24,26],through:[0,6,10,18,23,27,28,29],thu:18,tick:11,time:[0,1,6,7,11,13,14,15,16,17,21,22,23,26,27,29],timeout:22,timestamp:[15,22,25],tkagg:6,tmp:[13,14,15,24,26,29],token:[8,10,24],too:[24,29],tool:0,top:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29],total:22,trace:[0,15,20,22],track:16,tradit:28,transcrib:20,transduc:[0,15,16,24],transmiss:[15,21,22,25,26],tri:0,trigger:[0,1,3,6,13,16,19,26,27,28],tripl:[13,29],troubl:[18,27],troubleshoot:[16,27],truncat:[8,10,15,24],turn:[12,20,28,29],tutori:[14,16,18,23],tweet:[10,16,26],tweet_imag:[10,24],tweeter:[10,15,24,28],twitter:[0,13,15,16,26],two:[15,24,28],txt:[20,27],type:[6,9,11,12,15,16,17,18,20,21,22,23,24,26,27],typic:[24,27,29],ubuntu:14,udp:[0,5,7,16,20,22,23,24,25,27,29],unam:18,uncheck:24,unconsolid:24,under:29,understand:[24,28],unfamiliar:28,uniqu:0,unit:[14,15,24,26],univers:28,unix:[4,13,14,16,18,22,24],unless:24,unlimit:22,unmaintain:0,unmodifi:24,unpredict:0,unstabl:29,until:[9,15,22,25,28,29],unzip:18,updat:[6,11,22,24],update_plot:6,update_stream:22,upload:6,upon:[16,28],upper:15,usag:[20,22,27],use:[0,6,10,15,18,22,23,27,28],used:[0,3,6,8,10,13,15,17,19,20,22,24,27,28],useful:[0,15,17,27,28],user:[0,4,14,15,16,22,23,24,28,29],usernam:24,uses:[0,1,15,17,28,29],usestim:17,usg:10,using:[4,10,13,14,17,18,20,21,22,24,26,27,29],usual:[0,27,29],utc:[11,24],utcdatetim:[1,6,15,22],utcdatetime1:13,utcdatetime2:13,utcdatetime3:13,utcdatetime4:13,util:[17,20,24],valid:[15,24],valu:[12,18,21,22,24,26,28,29],vari:10,variabl:[11,12,15,17,18,22,26],varieti:20,variou:[3,6,16,18,19,27],vel:[7,24],veloc:[15,24],verbos:15,veri:[22,24],version:[4,5,7,8,9,10,12,15,18,20,24,25,26,28],vertic:[6,15,24],via:[0,8,15,28,29],vibrat:0,video:[16,18,23],virtual:18,visual:16,voltag:24,wai:[0,6,13,16,24,26,27,28,29],wait:[14,15,22,24],walkthrough:16,want:[4,13,14,15,24],wantedbi:14,warm:24,warn:[0,4,8,17,24],warnformat:17,warranti:[0,24],wast:26,watch:[27,28],web:24,websit:24,week:0,weird:6,welcom:[13,16],well:[13,14,24,29],were:[24,27],what:[4,14,16,24,27,28,29],whatev:[24,28],when:[0,1,2,6,10,11,13,15,16,18,21,22,24,25,28,30],whenev:24,where:[6,15,18,20,23,24,28,29],wherea:14,whether:[0,1,5,6,8,10,11,12,14,15,17,18,19,24,26,27],which:[0,8,14,15,17,19,20,21,22,24,27,28,29],white:15,why:[16,29],wifi:29,williamstown:24,win:[18,23],win_overrid:[4,24],win_ovr:4,window:[4,6,7,13,14,15,16,17],wish:[18,24,26,28],within:[14,20],without:[0,6,13,24,27,29],work:[13,14,15,18,22,24,28,29],worker:28,would:[15,18,24,26,27,28,29],write:[16,17,20,26],writer:16,written:[16,24],wrong:29,x86:18,x86_64:18,x_alarm:[9,26],x_data:26,x_imgpath:[9,26],x_masterqueu:26,x_packet:26,x_process:[9,26],x_reset:[9,26],x_send:26,x_term:[9,26],xyz:28,yellow:15,yet:0,yield:[17,24],you:[0,4,10,13,14,15,16,18,22,23,24,26,27,28,29],your:[4,10,13,14,16,21,22,23,24,26,29],your_modul:[26,28],your_shake_bot_id:24,your_test:26,yourself:[18,29],youtub:16,z0000:[22,24,26],zero:[15,22,29]},titles:["<span class=\"section-number\">1. </span>About rsudp","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_alert</span></code> (STA/LTA alarm)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_alertsound</span></code> (play sound)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_consumer</span></code> (master consumer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_custom</span></code> (run code)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_forward</span></code> (forward data)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_plot</span></code> (plot data)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_rsam</span></code> (RSAM analysis)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_telegram</span></code> (Telegram alerts)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_testing</span></code> (test consumer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_tweet</span></code> (Twitter alerts)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_write</span></code> (write data)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.client</span></code> (central module)","<span class=\"section-number\">2. </span>Contributing to this project","<span class=\"section-number\">5. </span>rsudp as a <code class=\"docutils literal notranslate\"><span class=\"pre\">systemctl</span></code> daemon","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.helpers</span></code> (helper functions)","<cite>rsudp</cite> 1.0.3","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.__init__</span></code> (initialization functions)","<span class=\"section-number\">2. </span>Installing rsudp","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.p_producer</span></code> (data producer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.packetize</span></code> (data formatter)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.packetloss</span></code> (track packets)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.raspberryshake</span></code> (main library)","<span class=\"section-number\">4. </span>Running rsudp","<span class=\"section-number\">3. </span>Modules and Settings","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.t_testdata</span></code> (test pre-producer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.test</span></code> (test helpers)","<span class=\"section-number\">7. </span>Testing and demoing rsudp","<span class=\"section-number\">1. </span>Program architecture and flow","<span class=\"section-number\">6. </span>Troubleshooting rsudp","YouTube walkthroughs"],titleterms:{"default":24,"function":[15,16,17,27],Adding:28,The:18,Using:[24,27],__init__:17,about:0,add:18,advanc:18,alarm:1,alarmsound:24,alert:[8,10,24],amplitud:24,analysi:7,app:24,architectur:28,bashrc:18,bot:24,broad:0,bug:13,c_alert:1,c_alertsound:2,c_consum:3,c_custom:4,c_forward:5,c_plot:6,c_rsam:7,c_telegram:8,c_test:9,c_tweet:10,c_write:11,can:[0,29],central:12,classroom:24,client:[12,28],code:[4,13,16,24],command:29,conda:29,consol:24,consum:[3,9,28],contribut:13,custom:24,daemon:14,data:[5,6,11,19,20,24,27,29],datacast:24,demo:[27,30],demonstr:[23,24],depend:24,detect:[24,30],document:16,dure:27,earthquak:[24,30],easi:18,exampl:24,exec:24,featur:28,file:28,firewal:29,flow:[27,28,29],formatt:20,forward:[5,24],gener:24,helper:[15,26],home:29,inconsist:29,index:16,initi:17,instal:[18,24],instruct:14,issu:29,layout:28,librari:22,line:18,linux:[18,24],live:[24,30],local:29,lta:[1,24],maco:[18,24],main:22,manual:23,master:3,messag:28,middl:29,minise:24,mode:23,modul:[12,24,27,28],network:29,notif:24,other:29,overview:0,own:[27,28],p_produc:19,packet:[20,21],packetloss:21,pass:28,plai:[2,24],plot:[6,24],pre:25,print:24,printdata:24,produc:[19,25,28],program:28,project:13,pydub:24,quiet:24,quit:23,raspberryshak:22,real:24,receiv:29,recommend:24,remot:29,restart:14,rsam:[7,24],rsudp:[0,1,2,3,4,5,6,7,8,9,10,11,12,14,15,16,17,18,19,20,21,22,23,25,26,27,29],run:[4,18,23,24,29],sampl:28,school:29,seismic:24,set:[24,27,28],setup:14,shake:29,side:29,sound:[2,24],sourc:18,special:0,sta:[1,24],start:23,stop:29,suggest:28,systemctl:14,t_testdata:25,telegram:[8,24],test:[9,23,25,26,27,28],thi:13,thread:28,time:24,track:21,trigger:24,troubleshoot:[14,29],tweet:24,twitter:[10,24],type:28,univers:29,unix:23,updat:18,upon:24,use:24,user:18,vault:24,wai:18,walkthrough:30,what:0,why:0,window:[18,23,24],write:[11,24],writer:24,your:[18,27,28],youtub:30}})
+Search.setIndex({docnames:["about","c_alert","c_alertsound","c_consumer","c_custom","c_forward","c_plot","c_rsam","c_telegram","c_testing","c_tweet","c_write","client","contributing","daemon","entrypoints","helpers","index","init","installing","p_producer","packetize","packetloss","raspberryshake","running","settings","t_testdata","test","testing","theory","troubleshooting","youtube"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":2,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.intersphinx":1,"sphinx.ext.todo":2,sphinx:56},filenames:["about.rst","c_alert.rst","c_alertsound.rst","c_consumer.rst","c_custom.rst","c_forward.rst","c_plot.rst","c_rsam.rst","c_telegram.rst","c_testing.rst","c_tweet.rst","c_write.rst","client.rst","contributing.rst","daemon.rst","entrypoints.rst","helpers.rst","index.rst","init.rst","installing.rst","p_producer.rst","packetize.rst","packetloss.rst","raspberryshake.rst","running.rst","settings.rst","t_testdata.rst","test.rst","testing.rst","theory.rst","troubleshooting.rst","youtube.rst"],objects:{"rsudp.__init__":{LevelFormatter:[18,1,1,""],add_debug_handler:[18,3,1,""],get_scap_dir:[18,3,1,""],init_dirs:[18,3,1,""],make_colors_friendly:[18,3,1,""],printE:[18,3,1,""],printM:[18,3,1,""],printW:[18,3,1,""],start_logging:[18,3,1,""]},"rsudp.__init__.LevelFormatter":{format:[18,2,1,""]},"rsudp.c_alert":{Alert:[1,1,1,""]},"rsudp.c_alert.Alert":{run:[1,2,1,""]},"rsudp.c_alertsound":{AlertSound:[2,1,1,""]},"rsudp.c_alertsound.AlertSound":{run:[2,2,1,""]},"rsudp.c_consumer":{Consumer:[3,1,1,""]},"rsudp.c_consumer.Consumer":{run:[3,2,1,""]},"rsudp.c_custom":{Custom:[4,1,1,""]},"rsudp.c_custom.Custom":{run:[4,2,1,""]},"rsudp.c_forward":{Forward:[5,1,1,""]},"rsudp.c_forward.Forward":{run:[5,2,1,""]},"rsudp.c_plot":{Plot:[6,1,1,""]},"rsudp.c_plot.Plot":{deconvolve:[6,2,1,""],figloop:[6,2,1,""],getq:[6,2,1,""],handle_close:[6,2,1,""],handle_resize:[6,2,1,""],mainloop:[6,2,1,""],qu:[6,2,1,""],run:[6,2,1,""],savefig:[6,2,1,""],set_sps:[6,2,1,""],setup_plot:[6,2,1,""],update_plot:[6,2,1,""]},"rsudp.c_rsam":{RSAM:[7,1,1,""]},"rsudp.c_rsam.RSAM":{run:[7,2,1,""]},"rsudp.c_telegram":{Telegrammer:[8,1,1,""]},"rsudp.c_telegram.Telegrammer":{run:[8,2,1,""]},"rsudp.c_testing":{Testing:[9,1,1,""]},"rsudp.c_testing.Testing":{run:[9,2,1,""]},"rsudp.c_tweet":{Tweeter:[10,1,1,""]},"rsudp.c_tweet.Tweeter":{run:[10,2,1,""]},"rsudp.c_write":{Write:[11,1,1,""]},"rsudp.c_write.Write":{elapse:[11,2,1,""],getq:[11,2,1,""],run:[11,2,1,""],set_sps:[11,2,1,""],slicestream:[11,2,1,""],write:[11,2,1,""]},"rsudp.client":{handler:[12,3,1,""],main:[12,3,1,""],mk_p:[12,3,1,""],mk_q:[12,3,1,""],run:[12,3,1,""],start:[12,3,1,""],test:[12,3,1,""],test_mode:[12,3,1,""]},"rsudp.entrypoints":{ep_cat_log:[15,3,1,""],ep_edit_settings:[15,3,1,""],ep_tailf_log:[15,3,1,""]},"rsudp.helpers":{conn_stats:[16,3,1,""],deconv_acc_inst:[16,3,1,""],deconv_rbm_inst:[16,3,1,""],deconv_vel_inst:[16,3,1,""],deconvolve:[16,3,1,""],default_settings:[16,3,1,""],dump_default:[16,3,1,""],fsec:[16,3,1,""],get_msg_path:[16,3,1,""],get_msg_time:[16,3,1,""],lesser_multiple:[16,3,1,""],msg_alarm:[16,3,1,""],msg_imgpath:[16,3,1,""],msg_reset:[16,3,1,""],msg_term:[16,3,1,""],read_settings:[16,3,1,""],resolve_extra_text:[16,3,1,""],set_channels:[16,3,1,""]},"rsudp.p_producer":{Producer:[20,1,1,""]},"rsudp.p_producer.Producer":{run:[20,2,1,""]},"rsudp.packetize":{cutoff_calc:[21,3,1,""],get_samps:[21,3,1,""],main:[21,3,1,""],packetize:[21,3,1,""]},"rsudp.packetloss":{main:[22,3,1,""],printTTLS:[22,3,1,""],run:[22,3,1,""],signal_handler:[22,3,1,""]},"rsudp.raspberryshake":{ConsumerThread:[23,1,1,""],copy:[23,3,1,""],getCHN:[23,3,1,""],getCHNS:[23,3,1,""],getDATA:[23,3,1,""],getSR:[23,3,1,""],getSTREAM:[23,3,1,""],getTIME:[23,3,1,""],getTR:[23,3,1,""],getTTLCHN:[23,3,1,""],get_inventory:[23,3,1,""],get_ip:[23,3,1,""],handler:[23,3,1,""],initRSlib:[23,3,1,""],make_trace:[23,3,1,""],openSOCK:[23,3,1,""],set_params:[23,3,1,""],update_stream:[23,3,1,""]},"rsudp.t_testdata":{TestData:[26,1,1,""]},"rsudp.t_testdata.TestData":{run:[26,2,1,""],send:[26,2,1,""]},"rsudp.test":{cancel_tests:[27,3,1,""],datadir_permissions:[27,3,1,""],is_connected:[27,3,1,""],logdir_permissions:[27,3,1,""],make_test_settings:[27,3,1,""],permissions:[27,3,1,""],ss_permissions:[27,3,1,""]},rsudp:{__init__:[18,0,0,"-"],c_alert:[1,0,0,"-"],c_alertsound:[2,0,0,"-"],c_consumer:[3,0,0,"-"],c_custom:[4,0,0,"-"],c_forward:[5,0,0,"-"],c_plot:[6,0,0,"-"],c_rsam:[7,0,0,"-"],c_telegram:[8,0,0,"-"],c_testing:[9,0,0,"-"],c_tweet:[10,0,0,"-"],c_write:[11,0,0,"-"],client:[12,0,0,"-"],entrypoints:[15,0,0,"-"],helpers:[16,0,0,"-"],p_producer:[20,0,0,"-"],packetize:[21,0,0,"-"],packetloss:[22,0,0,"-"],raspberryshake:[23,0,0,"-"],t_testdata:[26,0,0,"-"],test:[27,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:function"},terms:{"00z":23,"01t00":16,"10s":14,"15s":18,"16b":[0,25],"1dv7":23,"211704z":29,"21t19":23,"21t20":23,"23t06":29,"246777z":23,"292000z":23,"435214z":29,"532000z":23,"598944z":29,"599z":16,"boolean":[5,25,29],"break":26,"byte":[16,23,29],"case":[19,25,29,30],"class":[1,2,3,4,5,6,7,8,9,10,11,13,16,18,20,23,25,26,29],"default":[1,6,7,12,14,15,16,17,19,23,24,27,28,29],"final":[14,29,30],"float":[1,7,23,25],"function":[1,4,6,9,11,12,13,14,21,23,24,25,27,29],"import":[0,6,16,21,23,25,27,29],"int":[6,21,22,23,26],"long":[1,16,25],"new":[4,5,6,7,8,9,10,11,12,13,14,15,16,19,23,25,26,27,28,29,30],"public":[10,30],"return":[6,11,12,16,18,21,22,23,27,29],"short":[1,2,25],"super":29,"throw":25,"true":[1,4,6,7,11,12,16,18,22,23,25,27,29],"try":[16,27],"while":[0,25,29,30],Adding:[13,17,23],And:29,EHE:16,ENE:16,For:[9,12,16,19,23,25,26,28,29,30],One:30,Pis:25,That:[25,28],The:[0,1,3,5,6,10,11,12,13,15,16,17,18,20,21,22,23,24,25,27,28,29,30],Then:[26,27,29,30],There:[25,29,30],These:[9,15,16,18,25],Using:[17,29],Will:23,__init__:[17,29],_eq:[],abc:29,abil:[0,28],abl:[0,25,29,30],abort:30,about:[4,13,14,16,17,24,25,28,29,30],abov:[4,24,25,29,30],acc:[7,25],acceler:[16,25],acceleromet:[0,16,25],accept:16,access:[10,15,25],access_secret:[10,25],access_token:[10,25],access_token_secret:10,accommod:25,accompani:0,accordingli:25,account:[10,25],accumul:23,accur:16,achiev:[0,25],across:[0,6,25,30],action:[0,12,28,29],activ:[0,10,14,18,19,21,22,24,25,30],actual:[13,25],adapt:[18,23],add:[0,10,15,18,25,27,29],add_debug_handl:18,added:[13,15,23,29],adding:[15,28],addit:[0,8,10,13,16,18,29],addition:[0,25],addr:5,address:[5,7,16,23,25,26,30],adjust:[25,31],administr:19,advanc:15,advantag:25,adventur:13,after:[0,6,14,23,24,25,29,30],again:25,agenc:[10,25],agg:13,aggress:25,aim:0,alarm:[2,4,5,6,7,8,9,10,13,16,17,23,24,25,27,28,29],alarm_reset:[1,23],alarmsound:17,alert:[0,1,2,3,6,13,16,17,20,27,28,29],alertsound:[2,25,27,29],algorithm:[6,25],alia:15,alias:15,aliv:[9,20,23,26,29],all:[6,11,14,16,20,23,24,25,27,29,30],allow:[6,13,16,17,18,29,30],almost:30,along:29,alpha:13,alreadi:[13,14,18,25,29],also:[16,20,25,26,28,29,30],alt:25,although:[0,25,30],alwai:[13,14,25],amd:19,among:8,amount:25,amplitud:[7,17],anaconda3:19,anaconda:[19,24],analysi:[17,25],ani:[0,6,16,19,24,25,28,29,30],announc:18,anoth:[5,21,25,29,30],answer:[4,18,23,25],anyth:[12,28],anywai:25,api:[8,10,25,29],api_kei:25,api_secret:25,aplot:6,app:[10,13],appear:[25,30],append:[11,12,18,19,25],appli:[7,10,25],applic:0,appropri:[12,16,19,25],approxim:[8,25],apt:25,arbitrari:28,architectur:[17,19,28],archiv:28,arg:[1,7],argument:[21,22,23],armv7:19,armv7l:19,around:30,arriv:[2,6,25,26],artifici:28,ascii:21,asctim:18,ask:[19,30],aspect:0,assign:23,associ:18,assum:[0,25,29,30],atop:25,attach:[13,30],attribut:[18,23],audibl:0,audioseg:2,authent:25,automat:[8,19,23],avail:[0,16,17,19,23,24,25,27,28,31],averag:[1,2,25,30],avoid:[6,16,25],awai:0,axvlin:13,back:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30],backend:[6,13,27],backtick:[13,30],bak:19,bandpass:[1,25],bar:30,base:[16,19,21,25,28,29],bash:[14,15,19,24],bash_alias:15,basic:[25,28],bat:[19,24],batch:[13,24],beauti:0,becaus:[15,16,23,25],becom:28,bedrock:25,been:[0,15,19,25,30],beep:25,befor:[18,23,25,29],beforehand:25,begin:[6,24,26],behalf:10,behavior:6,being:[23,25,30],below:[0,17,19,25,28,29,30],beneath:[6,25],berryconda3:19,better:25,between:[22,23,25,29,30],beyond:13,bin:14,binari:25,bind:23,bit:28,block:6,blue:30,boaz:17,bonu:25,bool:[1,2,4,5,6,7,8,10,11,12,16,18,22,23,27,29],boom:[16,17,23,25,28,31],boot:14,bot:[8,10],botfath:25,both:[0,5,6,13,14,16,19,23,25,29],box:23,brace:29,brew:25,brief:[0,25],briefli:29,bring:[0,25],broad:17,broadcast:[6,8,25,29],bug:[17,23,30],build:[23,25,29],built:[5,10,25],button:30,buzzer:13,c_alarm:[],c_alert:[13,16,17,25],c_alertoff:27,c_alerton:27,c_alertsound:[17,25,29],c_anotheron:29,c_consum:[1,2,4,5,6,7,8,9,10,11,12,17,20,26,29],c_custom:[13,17,25,27],c_data:[],c_forward:[17,25,27],c_img:[9,27],c_imgpath:[],c_minise:27,c_mymodul:29,c_mytest:29,c_plai:27,c_plot:[16,17,25,27,29],c_print:27,c_printraw:25,c_process:[],c_reset:[],c_rsam:[17,27],c_telegram:[10,16,17,25,27,29],c_telegramimg:27,c_term:[],c_test:17,c_tweet:[17,25,27,29],c_tweetimg:27,c_write:[17,25,27],calcul:[0,1,3,6,16,20,21,23,25],calibr:16,call:[6,10,15,16,18,21,23,25,29,30],can:[4,5,6,10,13,14,16,17,19,21,23,24,25,28,29],cancel:27,cancel_test:27,cannot:[6,15,23,25,30],canva:6,capabl:[0,25,28,29],caption:29,captur:18,car:0,care:25,carri:18,cast:[0,30],cat:15,cat_log:[],caus:[6,11,19,23,25],centerpiec:0,central:[6,16,17],certain:[4,6,25],cha:[1,5,6,7,11,16],chan:[7,22,25,27],chang:[14,19,20,21,24,25,28,30],channel:[0,1,5,6,7,11,16,19,22,23,25,27],charact:[0,4,8,10,16,25],chart:[28,29],chat:[8,25],chat_id:[8,25],check:[4,14,16,17,19,25,28,30],child:29,chime:25,chn:23,choos:[5,19,24],chosen:[25,28],citizen:[16,23],classroom:28,clear:6,clearli:[13,30],click:[19,24,30],client:[2,17,23,24,28],clone:14,close:[6,12,24,29],closer:25,code:[0,6,14,19,23,27,29,30],codebas:0,codefil:[4,7,25],collect:0,color:[13,18],com:[13,14,17],combin:29,come:[10,23,24,25,30,31],comma:[25,29],command:[12,13,14,15,18,19,21,22,23,25,28],comment:[6,13],common:[23,30],commun:[0,13,25,29,30],comparison:25,compat:6,complex:0,comput:[0,18,19,24,30],conda:[19,21,22,24,28],config:[14,15,19,24,25,30],configur:[12,24,25,30],conform:29,confus:16,conn_stat:16,connect:[14,16,27,30],consecut:[22,23],consist:19,consol:[1,7,15,16,17],constant:22,constraint:25,construct:[16,29],consult:25,consum:[1,2,4,5,6,7,8,10,11,12,13,16,17,20,21,23,25,26,27],consumer_kei:10,consumer_secret:10,consumerthread:[23,29],contact:[0,25],contain:[0,6,13,16,21,23,25,26],content:[14,15,28],continu:[0,17,25,26,28,29],contribut:[0,17],control:[13,24,25,29],conveni:[17,25],convert:[4,16,21,23,25,28],cool:29,cooper:30,coordin:11,copi:[13,19,23,30],core:[1,6,11,16,21,23,27,28,29],correct:[4,23,25,30],correctli:[14,19,25,29],cost:0,could:[16,19,25,28,29,30],count:[16,23,25],counter:6,coupl:[18,25],cover:[19,24],cpu:23,creat:[13,14,16,17,18,19,23,24,25,28,29],creation:[8,13,14],crust:0,cryptographi:19,csv:[7,25],ctrl:[15,22,24,25,30],current:[9,16,19,23,27,29],custom:[4,13,17,27,28],custom_set:28,customiz:17,cut:16,cutoff_calc:21,d_matplotlib:27,d_pydub:27,daemon:[0,12,17],dai:11,daili:11,data:[0,1,2,3,4,7,8,9,10,16,17,22,23,24,26,27,29,31],data_dir:11,data_fil:26,datacast:[5,17,23,24,30],datadir_permiss:27,date:[11,18],datefmt:18,datetim:[16,18],dead:14,debain:14,debian:25,debug:[1,7,11,12,25,27,29],decim:23,decod:16,deconv:[1,6,7],deconv_acc_inst:16,deconv_rbm_inst:16,deconv_vel_inst:16,deconvolut:16,deconvolv:[6,16,25,27],def:29,default_set:16,defin:25,definit:[15,19],deliv:5,deliveri:30,demand:0,demo:[17,24],demonstr:[17,28],denot:[22,29],depend:[0,10,16,18,19,27,29],depict:28,deprec:25,depth:17,deriv:23,describ:[13,29,30],descript:[14,25,27,29],design:[0,25,26,28,30],desktop:19,destin:[3,5,12,25,29],detail:[0,6,10,23,25],detect:[0,17,24,27,28],detector:25,determin:[1,18,21,26],develop:[0,10,13,17,19,23,25,29],devic:[0,23,30],diagnost:14,diagram:[9,26,28],dict:[12,16,18,27],dictionari:[12,16,18,27,29],differ:[15,18,24,25,28,29,30],difficult:[10,15,30],digit:25,dir:25,directli:[0,23,25,29],directori:[14,18,19,25,27,28,29],disabl:[6,25],disclaim:[0,4,25],discov:28,disk:[11,16,26,28,29],disp:[7,25],displac:25,displai:[0,1,6,11,16,18,25,30],distanc:30,distribut:[0,3,5,10,19,20,25,29],docstr:29,document:[14,28,29],doe:[0,14,16,25,26,28],doesn:[16,19,25],doing:[25,28],don:[27,28,30],done:[14,16,19,21,23,28,29],doorbel:25,doubl:[19,24],down:[10,11,23],download:[19,23,25,27],downward:25,dport:23,draw:6,drawn:6,drop:[25,30],duckduckgo:30,due:[0,30],dump:[16,24,28],dump_default:16,duplic:29,durat:[1,2,6,25,27],dure:[9,17,26,29],dynam:6,each:[3,6,15,20,21,23,25,29,30],earli:[0,6,8,10,25],earlier:29,earth:0,earthquak:[0,17,24,28],easi:[0,13,24,25,29],easier:[10,13,25],easiest:27,easili:[13,23],echo:[15,19],edg:25,edit:[15,19,25,28],edit_set:25,editor:[15,25],educ:0,eew:[0,25],effect:6,effici:[13,29],ehn:16,ehz:[16,23,25,29],either:[5,6,7,21,24,25,29,30],elaps:[6,11,22],els:[12,30],elsewher:[0,14,25],embed:16,empti:25,emsc:10,enabl:[0,6,7,14,25,27,29],encod:23,encourag:29,end:[4,19,25,26,29],endpoint:25,endtest:26,enhanc:13,enn:16,enough:[0,23],ensur:25,enter:[14,19,24,25,28],entir:[14,25],entri:25,entrypoint:[14,17,25],environ:[19,21,22,24,28],enz:[16,25],ep_cat_log:15,ep_edit_set:15,ep_tailf_log:[14,15],eq_screenshot:[25,27],equal:25,equival:[14,15],errformat:18,error:[0,6,13,14,16,18,23,25,30],especi:[6,25],essenti:[28,29],etc:[13,19],ethernet:30,even:[13,25,28,30],event:[0,6,12,13,16,18,25],event_tim:[6,16],event_time_str:6,eventu:29,everi:[6,7,14,22,25,30],evt:6,exampl:[3,16,18,19,20,22,23,29,30],exceed:[1,25],except:[18,23,29],excess:25,exec:4,execreload:14,execstart:14,execstartpr:14,execstop:14,execut:[3,4,7,14,19,20,23,25,27,29],exist:[18,25,27,29,30],exit:[25,26,29],experi:[25,30],explain:25,explan:19,explicit:[0,25],explor:0,extens:0,extern:30,extra:[10,25],extra_text:[8,10,16,25],extract:23,extrem:[0,25],face:30,fact:0,fail:14,failur:29,fairli:30,fals:[0,1,2,3,4,5,6,7,8,9,10,11,12,16,18,20,21,22,23,25,26,27,29],far:[0,30],fast:[23,25],fdsn:[23,25,27],featur:[0,13,17,24,30],feed:28,feel:[13,17,29],fetch:9,few:25,ffmpeg:[19,25,29],field:[16,19,25],fifo:29,figloop:6,fignam:16,figur:[0,6,16,25,28,29,30],file:[4,5,6,10,11,14,15,16,18,19,21,24,25,26,27,28,30],filenam:6,filesystem:25,fill:25,filter:[1,25],find:[5,13,17,23,25,28],fine:23,firewal:17,first:[6,10,14,19,24,25,26,29,30],fit:16,fix:[19,30],flag:[1,20,28,29],flavor:25,flow:[9,17,24,26],fmt:18,focu:24,folder:[19,25],follow:[4,14,15,16,17,19,22,25,27,28,29,30],footstep:25,forc:[24,25],forese:21,forg:19,fork:25,form:[23,29],formal:29,format:[0,1,4,6,7,13,16,18,21,25,26,28,29,30],formatexcept:18,formatt:[17,29],formattim:18,formerli:24,forum:[13,29,30],forward:[7,11,17,24,27,30],found:25,four:[9,25],fraction:16,frame:[12,22,23],free:[8,25,29,30],freez:14,freq:16,frequenc:[21,22,23,25],friendli:18,from:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31],front:25,fsec:16,full:[6,16],fulli:[0,23,25],fullscreen:[6,25],fun:13,further:[0,25],futur:[19,21,25],fwaddr:[7,25],fwd_alarm:[5,25],fwd_data:[5,25],fwformat:[7,25],fwport:[7,25],gather:29,gempa:23,gener:[10,17,19,28,29,30],geophon:[0,6,16,25],get:[5,6,13,14,15,16,17,23,25,27,28,29,30],get_inventori:23,get_ip:23,get_msg_path:16,get_msg_tim:16,get_samp:21,get_scap_dir:18,getchn:23,getdata:23,getmessag:18,getq:[6,11,29],getsr:23,getstream:23,gettim:23,gettr:23,getttlchn:23,git:[14,19,24],github:[13,14,17,19,30],give:[10,25,30],given:[16,29,30],glad:30,global:[12,23],globe:30,goe:[25,29,30],going:[0,30],gone:10,got:29,gpio:13,gracefulli:20,grade:[0,25],graphic:14,grasp:0,grav:7,great:30,green:14,grei:14,ground:[3,20],group:25,guarante:[0,15,25,30],gui:[0,6,25],guid:[17,25],guidanc:12,hack:23,had:25,handl:[6,12,24],handle_clos:6,handle_res:6,handler:[12,18,22,23],happen:[6,25,29,30],hard:30,hardcod:21,harmless:19,has:[0,4,6,9,10,11,15,19,22,25,29,30],have:[0,4,6,13,14,16,19,23,25,28,29,30],hdf:[16,23,25],head:19,headless:0,hear:13,heart:6,help:[12,23,25,30],helper:[17,23],here:[0,1,4,13,18,25,27,29],hierarchi:[9,26,28,29],high:25,higher:25,highli:17,highpass:[1,25,27],hit:19,hobbi:[0,25],hold:25,home:[14,16,18,19,24,25,29],homebrew:25,host:[23,30],hostnam:27,hour:[0,10,22,30],how:[6,13,19,24,25,28,29],howev:[10,19,25,28,30],http:[13,14,17,29,30],hundredth:16,hyperlink:25,ian:17,iannesbitt:[14,16],idea:29,identifi:16,ifttt:13,ignor:[5,11,24,25,29],imag:[6,8,10,16,25,27,29],imgpath:[6,8,9,10,16,27,29],immedi:26,implement:29,impli:[0,25],implic:25,importerror:6,inabl:30,inact:14,includ:[0,10,14,23,24,25,28,29],incom:[1,24],incorpor:16,increas:23,increment:6,indefinit:30,independ:29,index:[1,2,3,4,5,6,7,8,9,10,11,12,20,21,22,23,25,26,27],indic:[12,13,20,30],individu:[28,30],inet:27,inf:21,infinit:29,info:[18,30],infoformat:18,inform:[0,10,16,18,23,25,29,30],ingest:30,init:[18,30],init_dir:18,initd:23,initi:[16,17,22,23,27,29],initrslib:23,input:[21,25],insert:25,inset:29,insid:[25,30],instal:[0,14,17,21,22,24,28,31],instanc:[16,29],instanti:29,instantli:29,instead:[10,28],institut:23,instruct:[4,17,19,25,27,28],instrument:[0,16,17,23,25],integ:[16,19,23],integr:[13,28,29],intend:[0,23,25],interact:13,interfac:18,intern:[25,30],internet:[27,28,30],interrupt:[14,15],interv:[7,25,27],intervent:0,intuit:0,inv:[16,23],inventori:[9,16,23,25,27],involv:13,ioerror:[23,30],is_connect:27,isn:13,isp:30,issu:[0,13,17,25,28,29],item:[25,29],iter:21,its:[0,9,10,14,16,23,25,27,28,29],itself:28,job:30,json:[5,7,15,16,19,24,25,28,29,30],jump:25,just:[9,11,25,29,30],keep:[0,23,25],kei:[10,19,25,30],keyboard:15,keystrok:22,kid:13,kill:14,kiosk:[0,6,25],know:[17,19,29,30],known:[24,25,30],kwarg:[1,7,23],label:[18,30],laid:29,laptop:19,larg:28,last:[11,25,26,28,29],later:[5,10,19],latest:[19,26],layout:17,lead:25,lean:0,learn:[0,13],least:[16,23,25],leav:[0,19],length:[16,25],less:16,lesser_multipl:16,let:[6,25,29,30],level:18,level_fmt:18,levelformatt:18,liabil:[0,25],libav:[25,29],librari:[6,16,17,24],licens:[0,25],life:[0,25],light:13,like:[3,9,13,14,15,18,19,20,23,24,25,28,29,30],likewis:1,limit:[0,8,10,16,25],line:[4,12,13,14,18,21,22,23,24,25,26,28,29,30],linecach:6,linestyl:13,linewidth:13,linger:14,link:25,linux:[14,15,17],list:[1,3,5,6,11,12,16,20,23,25,29],listen:[1,7,22,23,25,30],lite:[7,25],littl:6,live:[1,7,17],load:[12,25],loc:16,local:[0,17,22,23,25,28],localhost:[23,26],locat:[16,19,23,24,25,30],lock:25,log:[13,14,15,18,23,25,27,30],log_nam:18,logdir:27,logdir_permiss:27,logic:29,loginctl:14,lognam:[],logrecord:18,longer:[6,8,10,16,25],look:[0,10,14,15,17,19,20,23,25,28,29,30],loop:[6,23,26,29],lose:25,loss:[22,30],lost:22,lot:[6,29],low:[0,23],lowpass:[1,25,27],lta:17,lxml:19,mac:[19,25],machin:[14,23,25],maco:[14,15,17],made:[0,11,29],mai:[0,3,6,10,14,15,19,20,24,25,29,30],main:[3,6,12,17,21,22,25,29],mainloop:6,mainpid:14,maintain:[0,29],make:[0,4,6,10,12,13,18,19,21,23,25,28,29,30],make_colors_friendli:18,make_test_set:[27,28,29],make_trac:23,manag:25,mani:[0,6,10,25,29,30],manual:[17,19,25,30],master:[6,12,17,20,27,29],match:[16,25],matplotlib:[6,13,19,27,29],matter:25,max:1,max_len:16,mean:[0,4,19,21,25,28,29],meant:[0,25,28],measur:[7,16,25],media:[0,10,13,29],memori:[0,23],merg:23,messag:[1,2,3,4,5,6,7,8,9,10,11,13,15,16,17,18,20,21,25,26,27],method:[19,24,25],metric:16,mexican:[8,25],microsecond:16,middl:[17,29],might:[14,25,30],millisecond:[22,23],miniconda3:19,miniconda:19,minimum:0,minise:[11,17,21,27,28],minut:19,misconfigur:14,miss:25,mk_p:[12,29],mk_q:[12,29],mkdir:14,mode:[6,12,17,25,28],model:25,modif:23,modifi:[19,23,24,27,29],modul:[0,4,5,6,8,10,13,17,19,21,23,26,27],monei:0,monitor:[0,13,14,15,17,20,29,30],month:0,more:[4,5,6,10,13,15,18,23,24,25,28,29,30],most:[0,6,13,23,25,30],motion:[0,3,17,20,25],motor:13,move:[0,25,29,30],mp3:25,mp3file:25,mpl:27,mseed:[21,28],msg:[16,18],msg_alarm:16,msg_imgpath:16,msg_reset:16,msg_term:16,much:[0,15,25,29,30],multi:10,multipl:[0,5,15,16,25,29],must:[0,6,10,16,19,21,23,25,28,29,30],mymod:29,mymodul:29,n_internet:27,n_inventori:[9,27],n_port:27,nail:10,name:[6,11,16,18,19,22,23,24,25,27,28,29],nano:[14,15],nation:25,nativ:25,natur:0,navig:30,nearbi:25,nearest:16,nearli:0,necessari:[14,16,18,23],necessarili:25,need:[0,13,14,16,17,19,23,24,25,27,28,29,30],neg:[0,25],nesbitt:17,network:[5,14,23,25,26,27],never:25,newbot:25,newlin:[4,25],next:[14,19,24,25,26,29],nodata:23,node:[25,30],noder:25,nois:25,non:[0,25],none:[4,12,16,18],nonetyp:16,normal:[9,28,29],notabl:25,note:[0,28,30],notif:[0,8,10,17],notifi:[1,17,30],now:[6,14,19,23,25,29],nth:6,num:5,number:[6,8,16,21,22,23,25,29],numpi:19,object:[1,2,3,5,6,16,19,20,21,23,29],obspi:[1,3,6,11,16,19,20,21,23,25,28,29],occasion:[0,25],occur:[25,29],odir:18,off:[12,13,16,23,27,29],off_ev:13,offic:31,offici:[0,25],often:[25,30],old:[19,23],on_ev:13,onc:[6,10,25,28,29],one:[0,4,5,13,16,23,24,25,28,29,30],ones:13,onli:[0,8,9,10,12,15,16,23,24,25,29],onlin:30,open:[0,13,15,19,23,24,25,26,29],opensock:23,oper:[9,11,15,18,19,24,28],operand:18,option:[5,7,10,17,23,25,29],order:[0,5,6,14,16,17,19,23,25,29],ordinari:28,ordinarili:28,org:[13,29,30],organ:[17,29],orig:23,origin:[16,21],oserror:23,other:[0,1,3,5,6,8,11,13,14,17,19,20,24,25,29],otherwis:[7,11,22,27],our:[13,17,19,24,29,30,31],out:[6,16,17,18,23,25,28,29,30],outcom:0,outf:21,outlin:29,output:[0,12,13,14,15,16,18,21,25,27,30],output_dir:[11,16,25],outsid:30,over:[1,11,19,21,23,28,30],overload:30,overrid:[15,19],overview:[17,25],overwrit:24,overwritten:19,own:[0,13,17,23,25,27],p_data_dir:27,p_log_dir:27,p_log_fil:27,p_log_std:27,p_output_dir:27,p_produc:[1,3,16,17,23,25,29],p_screenshot_dir:27,packag:[21,25],packet:[5,6,7,9,11,16,17,23,25,26,27,28,30],packetloss:[17,30],pad:16,page:[1,2,3,4,5,6,7,8,9,10,11,12,13,17,20,21,22,23,25,26,27,28,30],panama:31,panamaigc:8,param:29,paramet:[1,2,3,4,5,6,7,8,9,10,11,12,16,17,18,20,21,22,23,25,26,27,29],parent:29,pars:[17,23,29],parser:29,part:[0,8,10,16,25,29],partial:16,partwai:19,pass:[2,3,4,5,6,9,12,13,16,17,20,21,22,23,25,26,27],past:[13,14,30],path:[14,15,16,18,19,24,25,27],patienc:30,paus:19,per:[6,11,16,21,23,24,25],perform:[0,25],perhap:[0,28],period:[0,22,25,30],perman:14,permiss:[10,27,28,30],person:17,piec:[0,25,28],pin:13,ping:25,pip:19,place:[15,21,23,25,29],plai:[13,17,19,27,28,29],plan:29,platform:[0,10,29],playback:25,pleas:[0,4,13,19,25,28,30],plenti:23,plot:[0,3,13,16,17,20,24,27,29],plt:13,plu:30,png:[16,25,29],point:[23,25,29],popul:[3,6,20,25],port:[5,7,16,20,21,22,23,24,25,26,27,28,29,30],portion:25,posit:[0,9,25,26],posix:15,post:[8,10,16,25,29,30],potenti:13,power:[0,6,9,23,25],pre:17,precis:[16,23],prefer:[15,17,24],prepar:26,preparatori:18,press:[24,25,30],pressur:[0,16,17,25],pretti:29,previou:[19,25],print:[7,15,16,17,18,22,23,27],printdata:[17,27],printfreq:22,printm:[18,23,29],printraw:25,printttl:22,printw:18,prior:[4,23,25,27],privat:30,privileg:19,probabl:[19,29],problem:[13,19,25,28,30],proc:12,proce:[19,25],process:[0,3,6,9,10,12,17,22,24,25,27,28,29],processor:[6,19,25],produc:[1,3,12,13,16,17,21,23,25,28],profession:[0,25],profil:[19,25],program:[0,11,13,14,17,19,22,23,24,25,26,27,28],programmat:29,project:[0,17],prompt:[15,19,24],proper:15,properli:6,properti:[0,25],protect:0,protocol:30,provid:[0,13,14,15,25,30],pull:[19,29],purpos:[11,25,29],put:[6,16,20,25],pydub:[2,27],pyplot:[13,29],python3:21,python:[0,4,16,17,19,21,25,29],qt5:6,quak:25,quarter:6,question:[4,11,13,21,25,29],queue:[1,2,3,4,5,6,7,8,9,10,11,12,13,16,20,25,26,27,29],quickli:[0,8,10],quiet:[7,27],quit:[2,4,6,7,17,20,26,29],r24fa:[16,27,29],r3bcf:23,r940d:23,rais:[1,6,23],ram:0,raspberri:[0,10,14,16,17,19,21,23,25,26,27,28,30,31],raspberryshak:[13,16,17,29,30],raspbian:[14,19,25],raspishak:[13,14,17],rate:[6,10,16,22,23,26,28],ratio:[1,25],raw:[5,16,25],rcb43:23,reach:26,read:[1,2,3,4,5,7,8,10,11,16,17,19,21,23,25,26,28,29],read_set:16,reader:26,readi:[16,19,24,25],readili:25,real:[0,7,14,17],realli:25,reason:[14,19,24,30],reboot:14,rebuild:24,receiv:[15,20,23,25,27,28,29],recept:9,record:[0,6,16,18,28],recur:30,recurs:[1,25],recursive_sta_lta:[1,25],red:[14,29],redraw:6,referenti:23,reflect:[16,30],refresh:[6,25],regard:[0,14],regist:25,reject:25,rel:10,relat:6,releas:[19,25],relev:[25,27,29,30],reli:[0,25,29],reliabl:[10,23],reload:[14,15],remain:0,remind:29,remot:[17,28],remov:25,render:25,replac:19,report:[12,16,22,23,29,30],repositori:[0,19,24,25],repres:9,request:[23,29],requir:[0,4,23,25,28,29],reset:[1,5,9,13,16,23,25,27,29],reset_tim:16,resid:30,resiz:6,resolv:[23,25],resolve_extra_text:16,resourc:[13,17],respect:[25,29],respons:[25,30],rest:[28,29],restart:[0,17,25,30],restartsec:14,restructuredtext:29,result:[12,16,25,27],rewritten:19,richard:17,right:[16,23,29,30],rigor:10,room:0,round:16,router:30,routin:[0,6,11,12,13,28],row:[29,30],rs1d:21,rs4d:[16,25],rsam:[17,27],rsstn:23,rsudp:[13,25,29,31],rsudp_set:[15,19,24,25,30],rtype:29,rule:10,run:[0,1,2,3,5,6,7,8,9,10,11,12,13,14,15,17,20,21,22,23,26,27,28,29,31],runtim:29,s0cde:23,safer:30,sai:[14,25,30],said:25,same:[23,25,26,28,30],sampl:[6,11,16,21,23,28],sasmex:[8,25],save:[0,6,16,18,25,29],save_pct:6,savefig:[6,29],scenario:25,scienc:[0,16,23],scipi:19,screen:[18,25],screencap:6,screenshot:[6,13,16,18,25,27,29],script:[13,14,19,23,24,25,28,30],scroll:30,search:[1,2,3,4,5,6,7,8,9,10,11,12,17,20,21,22,23,25,26,27,30],sec:16,second:[1,2,6,7,11,14,16,19,22,23,25,28,29,30],secret:[10,25],section:[0,5,19,25,29,30],secur:30,sediment:25,see:[0,1,2,4,5,6,7,8,9,10,12,13,14,17,18,20,21,23,24,25,26,28,29,30],seek:26,seem:0,segment:11,seiscomp3:23,seismic:[0,7,17,21,23,28],seismogram:[6,25],seismograph:[0,17],seismolog:0,self:[2,4,6,7,9,11,16,23,26,29],send:[1,6,7,8,10,13,16,18,23,24,25,26,27,28,29,30],send_imag:[8,25],sender:[8,16,18,23,29],sensor:21,sent:[1,2,3,4,5,6,7,8,9,10,11,21,23,25,26,29,30],separ:[6,25],sequenc:[25,29],serial:16,server:[25,27],servic:[6,8,10,14,25,27,30],set:[0,1,4,5,6,7,10,11,12,13,14,15,16,17,19,22,23,24,26,27,30,31],set_channel:16,set_param:23,set_sp:[6,11],settings_loc:16,setup:[12,15,17,23,30],setup_plot:6,seven:9,sever:[19,24,25,30],shake:[0,6,10,14,16,17,21,23,24,25,26,27,28,31],shorten:25,should:[0,1,12,14,15,19,20,23,25,29,30],show:[0,12,14,28,29],shown:29,shutdown:6,shz:[16,25],side:17,sig:12,sign:25,signal:[1,6,12,22,23,25,29],signal_handl:22,signum:[22,23],similar:[0,13,19,29],similarli:25,simpl:[0,10,11,16,25,26],simpli:[19,25],simultan:[0,25],sinc:[13,22,23,25,29,30],singl:[5,23,25],site:30,size:25,skip:19,sleep:14,slice:11,slicestream:11,slightli:[28,30],slow:30,slower:[6,25],small:[25,28,30],social:[0,10,13,29],socket:[23,26],softwar:[0,13,19,24,25,28,29],solut:23,solv:30,some:[6,13,14,15,16,19,23,24,25,27,29,30],somehow:14,someon:28,someth:[13,14,17,23,25,29,30],sometim:10,somewher:[25,30],sonar:25,soon:31,sort:29,sound:[0,1,13,17,19,24,27,28,29],soundloc:2,sourc:[0,14,16,24,25,29,30],space:[18,25],span:10,special:17,specif:[1,6,12,16,19,22,23,28,30],specifi:[5,7,16,18,23,25,26,27,28,29,30],spectrogram:[6,25],speed:[25,26],spent:0,sphinx:29,spin:29,sps:[16,23],sqlalchemi:19,ss_permiss:27,sta:[2,17],stabl:0,stackoverflow:[4,18,23,25],stamp:18,stand:29,standard:[0,4,25,29],start:[5,9,11,12,14,17,22,23,25,26,27,28,29],start_log:18,startup:6,stat:16,state:[0,12,13,23,24,25,27],station:[16,23,25,27],stationview:25,statist:22,statu:[14,20,29],stdout:[25,27],steim2:25,step:[18,19,25,29],still:[25,30],stn:23,stomp:25,stop:[12,20,24],store:[23,30],str:[1,4,5,6,7,8,10,11,16,18,21,23,26,27,29],straightforward:30,strang:16,stream:[1,3,6,9,11,16,17,20,21,22,23,24,25,29,31],street:0,strftime:[16,18],strict:30,stricter:10,string:[4,6,16,18,23,25,26,29],structur:[14,27],student:28,stuff:[6,29],style:[4,15,25],sub:[3,12,16,17,23,29],submit:[29,30],subsequ:25,substitut:[0,25],succe:25,success:27,successfulli:25,sudden:[0,17,25],sudo:[14,25],suffer:0,suffic:0,suffici:25,suggest:[13,17],suit:[8,25],summari:28,suppli:[12,25],support:[16,29,30],suppress:7,sure:[4,13,25,28,30],suspici:10,switcher:25,sys:29,system:[0,8,14,15,19,23,24,25],systemctl:[0,17],systemd:14,t_testdata:[17,28],tab:25,tag:13,tail:[14,15],tailf:[14,15],tailf_log:[],take:[6,12,19,25],taken:12,talk:30,target:[14,30],task:[0,3,20],task_don:29,team:[16,30],technic:30,telegram:[0,10,13,16,17,27],telegramm:[8,16,25,29],tell:[19,25,28,29,30],term:[1,2,4,6,7,9,16,20,25,26,27,29],termin:[19,24],test:[0,1,2,3,4,5,6,7,8,10,11,12,14,15,16,17,18,19,20,21,25],test_mod:12,testb:23,testdata:[9,26,28],testdir:27,text:[10,12,14,15,16,18,21,25,27,28,30],than:[5,6,10,14,16,19,24,25,28,30],thank:[29,30],thei:[13,18,20,23,24,27,28,29,30],them:[12,16,18,25,26,29,30],theoret:25,theori:[0,29],therefor:17,thi:[0,1,3,4,5,6,8,9,10,11,14,15,16,17,18,19,21,22,23,24,25,26,27,28,29,30],thing1:29,thing:[0,3,6,8,13,18,20,25,29,30],think:10,thoroughli:25,those:[3,5,21,23],thread:[5,9,12,17,20,23,25,26,28],threadsaf:[6,18],three:[25,29,30],thresh:1,threshold:[1,25,27],through:[0,6,10,19,24,28,29,30],thu:19,tick:11,time:[0,1,6,7,11,13,14,16,17,18,22,23,24,27,28,30],timeout:23,timestamp:[16,23,26],tkagg:6,tmp:[13,14,15,25,27,30],token:[8,10,25],too:[25,30],tool:0,top:[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30],total:23,trace:[0,16,21,23],track:17,tradit:29,transcrib:21,transduc:[0,16,17,25],transmiss:[16,22,23,26,27],tri:0,trigger:[0,1,3,6,13,17,20,27,28,29],tripl:[13,30],troubl:[19,28],troubleshoot:[17,28],truncat:[8,10,16,25],turn:[12,21,29,30],tutori:[14,17,19,24],tweet:[10,17,27],tweet_imag:[10,25],tweeter:[10,16,25,29],twitter:[0,13,16,17,27],two:[16,25,29],txt:[21,28],type:[6,9,11,12,16,17,18,19,21,22,23,24,25,27,28],typic:[25,28,30],ubuntu:14,udp:[0,5,7,17,21,23,24,25,26,28,30],unam:19,uncheck:25,unconsolid:25,under:30,understand:[25,29],unfamiliar:29,uniqu:0,unit:[14,16,25,27],univers:29,unix:[4,13,14,17,19,23,25],unless:25,unlimit:23,unmaintain:0,unmodifi:25,unpredict:0,unstabl:30,until:[9,15,16,23,26,29,30],unzip:19,updat:[6,11,23,25],update_plot:6,update_stream:23,upload:6,upon:[17,29],upper:16,usag:[21,23,28],use:[0,6,10,16,19,23,24,28,29],used:[0,3,6,8,10,13,16,18,20,21,23,25,28,29],useful:[0,16,18,28,29],user:[0,4,14,15,16,17,23,24,25,29,30],usernam:25,uses:[0,1,15,16,18,29,30],usestim:18,usg:10,using:[4,10,13,14,18,19,21,22,23,25,27,28,30],usual:[0,28,30],utc:[11,25],utcdatetim:[1,6,16,23],utcdatetime1:13,utcdatetime2:13,utcdatetime3:13,utcdatetime4:13,util:[18,21,25],valid:[16,25],valu:[12,19,22,23,25,27,29,30],vari:10,variabl:[11,12,16,18,19,23,27],varieti:21,variou:[3,6,17,19,20,28],vel:[7,25],veloc:[16,25],verbos:16,veri:[23,25],version:[4,5,7,8,9,10,12,15,16,19,21,25,26,27,29],vertic:[6,16,25],via:[0,8,15,29,30],vibrat:0,video:[17,19,24],virtual:19,visual:17,voltag:25,wai:[0,6,13,17,25,27,28,29,30],wait:[14,15,23,25],walkthrough:17,want:[4,13,14,16,25],wantedbi:14,warm:25,warn:[0,4,8,18,25],warnformat:18,warranti:[0,25],wast:27,watch:[28,29],web:25,websit:25,week:0,weird:6,welcom:[13,17],well:[13,14,25,30],were:[25,28],what:[4,14,17,25,28,29,30],whatev:[25,29],when:[0,1,2,6,10,11,13,16,17,19,22,23,25,26,29,31],whenev:25,where:[6,16,19,21,24,25,29,30],wherea:14,whether:[0,1,5,6,8,10,11,12,14,16,18,19,20,25,27,28],which:[0,8,14,16,18,20,21,22,23,25,28,29,30],white:16,why:[17,30],wifi:30,williamstown:25,win:[19,24],win_overrid:[4,25],win_ovr:4,window:[4,6,7,13,14,15,17,18],wish:[19,25,27,29],within:[14,21],without:[0,6,13,25,28,30],work:[13,14,16,19,23,25,29,30],worker:29,would:[16,19,25,27,28,29,30],write:[17,18,21,27],writer:17,written:[17,25],wrong:30,x86:19,x86_64:19,x_alarm:[9,27],x_data:27,x_imgpath:[9,27],x_masterqueu:27,x_packet:27,x_process:[9,27],x_reset:[9,27],x_send:27,x_term:[9,27],xyz:29,yellow:16,yet:0,yield:[18,25],you:[0,4,10,13,14,16,17,19,23,24,25,27,28,29,30],your:[4,10,13,14,17,22,23,24,25,27,30],your_modul:[27,29],your_shake_bot_id:25,your_test:27,yourself:[19,30],youtub:17,z0000:[23,25,27],zero:[16,23,30]},titles:["<span class=\"section-number\">1. </span>About rsudp","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_alert</span></code> (STA/LTA alarm)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_alertsound</span></code> (play sound)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_consumer</span></code> (master consumer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_custom</span></code> (run code)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_forward</span></code> (forward data)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_plot</span></code> (plot data)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_rsam</span></code> (RSAM analysis)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_telegram</span></code> (Telegram alerts)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_testing</span></code> (test consumer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_tweet</span></code> (Twitter alerts)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.c_write</span></code> (write data)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.client</span></code> (central module)","<span class=\"section-number\">2. </span>Contributing to this project","<span class=\"section-number\">5. </span>rsudp as a <code class=\"docutils literal notranslate\"><span class=\"pre\">systemctl</span></code> daemon","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.entrypoints</span></code> (convenience functions)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.helpers</span></code> (helper functions)","<cite>rsudp</cite> 1.0.3","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.__init__</span></code> (initialization functions)","<span class=\"section-number\">2. </span>Installing rsudp","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.p_producer</span></code> (data producer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.packetize</span></code> (data formatter)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.packetloss</span></code> (track packets)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.raspberryshake</span></code> (main library)","<span class=\"section-number\">4. </span>Running rsudp","<span class=\"section-number\">3. </span>Modules and Settings","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.t_testdata</span></code> (test pre-producer)","<code class=\"xref py py-data docutils literal notranslate\"><span class=\"pre\">rsudp.test</span></code> (test helpers)","<span class=\"section-number\">7. </span>Testing and demoing rsudp","<span class=\"section-number\">1. </span>Program architecture and flow","<span class=\"section-number\">6. </span>Troubleshooting rsudp","YouTube walkthroughs"],titleterms:{"default":25,"function":[15,16,17,18,28],Adding:29,The:19,Using:[25,28],__init__:18,about:0,add:19,advanc:19,alarm:1,alarmsound:25,alert:[8,10,25],amplitud:25,analysi:7,app:25,architectur:29,bashrc:19,bot:25,broad:0,bug:13,c_alert:1,c_alertsound:2,c_consum:3,c_custom:4,c_forward:5,c_plot:6,c_rsam:7,c_telegram:8,c_test:9,c_tweet:10,c_write:11,can:[0,30],central:12,classroom:25,client:[12,29],code:[4,13,17,25],command:30,conda:30,consol:25,consum:[3,9,29],contribut:13,conveni:15,custom:25,daemon:14,data:[5,6,11,20,21,25,28,30],datacast:25,demo:[28,31],demonstr:[24,25],depend:25,detect:[25,31],document:17,dure:28,earthquak:[25,31],easi:19,entrypoint:15,exampl:25,exec:25,featur:29,file:29,firewal:30,flow:[28,29,30],formatt:21,forward:[5,25],gener:25,helper:[16,27],home:30,inconsist:30,index:17,initi:18,instal:[19,25],instruct:14,issu:30,layout:29,librari:23,line:19,linux:[19,25],live:[25,31],local:30,lta:[1,25],maco:[19,25],main:23,manual:24,master:3,messag:29,middl:30,minise:25,mode:24,modul:[12,25,28,29],network:30,notif:25,other:30,overview:0,own:[28,29],p_produc:20,packet:[21,22],packetloss:22,pass:29,plai:[2,25],plot:[6,25],pre:26,print:25,printdata:25,produc:[20,26,29],program:29,project:13,pydub:25,quiet:25,quit:24,raspberryshak:23,real:25,receiv:30,recommend:25,remot:30,restart:14,rsam:[7,25],rsudp:[0,1,2,3,4,5,6,7,8,9,10,11,12,14,15,16,17,18,19,20,21,22,23,24,26,27,28,30],run:[4,19,24,25,30],sampl:29,school:30,seismic:25,set:[25,28,29],setup:14,shake:30,side:30,sound:[2,25],sourc:19,special:0,sta:[1,25],start:24,stop:30,suggest:29,systemctl:14,t_testdata:26,telegram:[8,25],test:[9,24,26,27,28,29],thi:13,thread:29,time:25,track:22,trigger:25,troubleshoot:[14,30],tweet:25,twitter:[10,25],type:29,univers:30,unix:24,updat:19,upon:25,use:25,user:19,vault:25,wai:19,walkthrough:31,what:0,why:0,window:[19,24,25],write:[11,25],writer:25,your:[19,28,29],youtub:31}})
