+++ docs/settings.html
@@ -152,6 +152,7 @@
 <li class="toctree-l1"><a class="reference internal" href="init.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.__init__</span></code> (initialization functions)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="raspberryshake.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.raspberryshake</span></code> (main library)</a></li>
 <li class="toctree-l1"><a class="reference internal" href="helpers.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.helpers</span></code> (helper functions)</a></li>
+<li class="toctree-l1"><a class="reference internal" href="entrypoints.html"><code class="xref py py-data docutils literal notranslate"><span class="pre">rsudp.entrypoints</span></code> (convenience functions)</a></li>
 </ul>
 <p class="caption"><span class="caption-text">Clients</span></p>
 <ul>
@@ -254,7 +255,7 @@ Additionally after changing settings, you will need to restart <code class="code
 where <code class="code bash docutils literal notranslate"><span class="name variable"><span class="pre">$HOME</span></span></code> is your home directory (often shortened to <code class="code bash docutils literal notranslate"><span class="pre">~/</span></code>).
 To output default settings to a different location, see <a class="reference internal" href="running.html#running-manually"><span class="std std-ref">Starting manually on Windows or Unix</span></a>.</p>
 <p>For convenience, you can use rsudp’s built-in <code class="code bash docutils literal notranslate"><span class="pre">rs-settings</span></code> command
-(<a class="reference internal" href="helpers.html#rsudp.helpers.edit_settings" title="rsudp.helpers.edit_settings"><code class="xref py py-func docutils literal notranslate"><span class="pre">rsudp.helpers.edit_settings()</span></code></a>) to call your system’s default editor to edit the settings
+(<code class="xref py py-func docutils literal notranslate"><span class="pre">rsudp.entrypoints.edit_settings()</span></code>) to call your system’s default editor to edit the settings
 file.</p>
 <div class="section" id="settings-general-settings">
 <h2><span class="section-number">3.1. </span><code class="code docutils literal notranslate"><span class="pre">settings</span></code> (general settings)<a class="headerlink" href="#settings-general-settings" title="Permalink to this headline">¶</a></h2>
