+++ docsrc/entrypoints.rst
@@ -1,0 +1,11 @@
+:py:data:`rsudp.entry_points` (convenience functions)
+=====================================================
+
+These are some convenience functions for editing rsudp's settings and
+monitoring log output.
+
+.. automodule:: rsudp.entry_points
+    :members:
+
+
+`Back to top ↑ <#top>`_
